Launch ablation() with the same command-line options as ablation.exe.

Launch ablation() with the same command-line options as ablation.exe.

ablation_cmdline(argv = commandArgs(trailingOnly = TRUE))

Arguments

argv

(character())
The arguments provided on the R command line as a character vector, e.g., c("-i", "irace.Rdata", "--src", 1).

Value

A list containing the following elements:

configurations

Configurations tested in the ablation.

instances

A matrix with the instances used in the experiments. First column has the instances IDs from iraceResults$scenario$instances, second column the seed assigned to the instance.

experiments

A matrix with the results of the experiments (columns are configurations, rows are instances).

scenario

Scenario object with the settings used for the experiments.

trajectory

IDs of the best configurations at each step of the ablation.

best

Best configuration found in the experiments.

Details

The function reads the parameters given on the command line used to invoke R, launches ablation() and possibly plotAblation(). List of command-line parameters.

## -i,--log-file            path to the (.Rdata) file created by irace from which
##                          the `iraceResults` object will be loaded
## -S,--src                 Source configuration ID
## -T,--target              Target configuration ID. By default the best
##                          configuration found by irace
## -P,--params              Specific parameter names to be used for the ablation
##                          (separated with commas). By default use all
## -t,--type                Type of ablation to perform: "full" will execute each
##                          configuration on all `--n-instances` to determine the
##                          best-performing one; "racing" will apply racing to
##                          find the best configurations
## -n,--n-instances         Number of instances to be used for the "full"
##                          ablation, if not provided firstTest instances are used
## -s,--seed                Integer value to use as seed for the random number
##                          generation
## -o,--output-file         Log file to save the ablation log. If "", the results
##                          are not saved to a file.
## -p,--plot                Output filename (.pdf) for the plot. If not given, no
##                          plot is created
## -O,--plot-type           Type of plot. Supported values are "mean" and
##                          "boxplot"

Author

Manuel López-Ibáñez

Examples

ablation_cmdline("--help")
#> #------------------------------------------------------------------------------
#> # irace: An implementation in R of (Elitist) Iterated Racing
#> # Version: 3.5.ec8110c-dirty
#> # Copyright (C) 2010-2020
#> # Manuel Lopez-Ibanez     <manuel.lopez-ibanez@manchester.ac.uk>
#> # Jeremie Dubois-Lacoste  
#> # Leslie Perez Caceres    <leslie.perez.caceres@ulb.ac.be>
#> #
#> # This is free software, and you are welcome to redistribute it under certain
#> # conditions.  See the GNU General Public License for details. There is NO
#> # WARRANTY; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
#> #
#> # irace builds upon previous code from the race package:
#> #     race: Racing methods for the selection of the best
#> #     Copyright (C) 2003 Mauro Birattari
#> #------------------------------------------------------------------------------
#> # installed at: /home/runner/work/_temp/Library/irace
#> # called with: --help
#> -i,--log-file            path to the (.Rdata) file created by irace from which
#>                          the `iraceResults` object will be loaded
#> -S,--src                 Source configuration ID
#> -T,--target              Target configuration ID. By default the best
#>                          configuration found by irace
#> -P,--params              Specific parameter names to be used for the ablation
#>                          (separated with commas). By default use all
#> -t,--type                Type of ablation to perform: "full" will execute each
#>                          configuration on all `--n-instances` to determine the
#>                          best-performing one; "racing" will apply racing to
#>                          find the best configurations
#> -n,--n-instances         Number of instances to be used for the "full"
#>                          ablation, if not provided firstTest instances are used
#> -s,--seed                Integer value to use as seed for the random number
#>                          generation
#> -o,--output-file         Log file to save the ablation log. If "", the results
#>                          are not saved to a file.
#> -p,--plot                Output filename (.pdf) for the plot. If not given, no
#>                          plot is created
#> -O,--plot-type           Type of plot. Supported values are "mean" and
#>                          "boxplot"