Generate a command-line representation of a configuration

buildCommandLine receives two vectors, one containing the values of the parameters, the other containing the switches of the parameters. It builds a string with the switches and the values that can be used as a command line to call the program to be tuned, thus generating one candidate configuration.

buildCommandLine(values, switches)

Arguments

values

A vector containing the value of each parameter for the candidate configuration.

switches

A vector containing the switches of each paramter (in an order that corresponds to the values vector).

Value

A string concatenating each element of switches and values for all parameters with a space between each pair of parameters (but none between the switches and the corresponding values).

Author

Manuel López-Ibáñez and Jérémie Dubois-Lacoste

Examples

switches <- c("--switch1 ", "--switch2 ")
values <- c("value_1", "value_2")
buildCommandLine (values, switches)
#> [1] " --switch1 value_1 --switch2 value_2"
## Build a command-line from the results produced by a previous run of irace.
# First, load the data produced by irace.
irace.logfile <- file.path(system.file(package="irace"),
                           "exdata", "irace-acotsp.Rdata")
load(irace.logfile)
allConfigurations <- iraceResults$allConfigurations
parameters <- iraceResults$parameters
apply(allConfigurations[1:10, unlist(parameters$names)], 1, buildCommandLine,
      unlist(parameters$switches))
#>                                                                                                                 1 
#>                                      " --as --localsearch 0 --alpha 1.0000 --beta 1.0000 --rho  0.9500 --ants 10" 
#>                                                                                                                 2 
#> " --eas --localsearch 1 --alpha 0.1468 --beta 4.9900 --rho  0.2708 --ants 22 --nnls 50 --dlb 1 --elitistants 674" 
#>                                                                                                                 3 
#>                   " --eas --localsearch 0 --alpha 2.9351 --beta 6.0060 --rho  0.1717 --ants 63 --elitistants 135" 
#>                                                                                                                 4 
#> " --eas --localsearch 2 --alpha 3.3636 --beta 2.8779 --rho  0.5877 --ants 10 --nnls  8 --dlb 1 --elitistants  50" 
#>                                                                                                                 5 
#>                                      " --as --localsearch 0 --alpha 1.7719 --beta 3.4517 --rho  0.4868 --ants 76" 
#>                                                                                                                 6 
#>                       " --ras --localsearch 0 --alpha 3.9847 --beta 4.9987 --rho  0.1967 --ants  7 --rasranks 91" 
#>                                                                                                                 7 
#> " --eas --localsearch 1 --alpha 1.4327 --beta 2.5215 --rho  0.2899 --ants 10 --nnls 38 --dlb 0 --elitistants 329" 
#>                                                                                                                 8 
#> " --eas --localsearch 1 --alpha 4.6157 --beta 8.1128 --rho  0.7505 --ants 73 --nnls 34 --dlb 0 --elitistants 638" 
#>                                                                                                                 9 
#>       " --acs --localsearch 2 --alpha 4.9484 --beta 3.1750 --rho  0.4944 --ants 13 --nnls 47 --q0 0.0998 --dlb 1" 
#>                                                                                                                10 
#>                    " --as --localsearch 1 --alpha 4.4563 --beta 0.2244 --rho  0.9942 --ants 52 --nnls  7 --dlb 1"