Test that the given irace scenario can be run.

Test that the given irace scenario can be run by checking the scenario settings provided and trying to run the target-algorithm.

Usage

checkIraceScenario(scenario)

Arguments

scenario

list()
Data structure containing irace settings. The data structure has to be the one returned by the function defaultScenario() or readScenario().

Value

returns TRUE if successful and gives an error and returns FALSE otherwise.

Details

If the parameters argument is missing, then the parameters will be read from the file parameterFile given by scenario. If parameters is provided, then parameterFile will not be read. This function will try to execute the target-algorithm.

See also

Author

Manuel López-Ibáñez and Jérémie Dubois-Lacoste