Return the elite configurations of the final iteration.

Return the elite configurations of the final iteration.

getFinalElites(
  iraceResults = NULL,
  logFile = NULL,
  n = 0,
  drop.metadata = FALSE
)

Arguments

iraceResults

Object created by irace and saved in scenario$logFile.

logFile

Log file created by irace, this file must contain the iraceResults object.

n

Number of elite configurations to return, if n is larger than the number of configurations, then only the existing ones are returned. The default (n=0) returns all of them.

drop.metadata

Remove metadata, such the configuration ID and the ID of the parent, from the returned configurations. See removeConfigurationsMetaData.

Value

A data frame containing the elite configurations required.

Author

Manuel López-Ibáñez and Leslie Pérez Cáceres

Examples

log_file <- system.file("exdata/irace-acotsp.Rdata", package="irace", mustWork=TRUE)
print(removeConfigurationsMetaData(getFinalElites(logFile=log_file, n=1)))
#>    algorithm localsearch  alpha  beta    rho ants nnls     q0 dlb rasrank
#> 99       acs           3 3.5404 3.176 0.7968   67   25 0.5218   1      NA
#>    elitistants
#> 99          NA