Command-line interface to irace2pyimp

This is a command-line interface for calling the irace2pyimp function, which converts an irace.Rdata file into the input format supported by the parameter importance analysis tool PyImp (https://github.com/automl/ParameterImportance).
The best way to use this command line interface is to run the script irace-to-pyimp.
To see usage of the script, run: irace-to-pyimp --help

irace2pyimp_cmdline(argv = commandArgs(trailingOnly = TRUE))

Arguments

argv
(character())
Command-line arguments.

Value

None.

See also

Examples

irace2pyimp_cmdline("--help")

#> #------------------------------------------------------------------------------
#> # irace2pyimp converts an irace.Rdata file into input format supported by the
#> # parameter importance analysis tool PyImp (https://github.com/automl/ParameterImportance)
#> #
#> # irace2pyimp was developed by Nguyen Dang <nttd@st-andrews.ac.uk>
#> #
#> # PyImp is developed by the AutoML-Freiburg-Hannover group (https://www.automl.org/)
#> #
#> # To see usage: irace2pyimp --help
#> # See examples in `$IRACE_HOME/inst/examples/irace2pyimp/acotsp/run.sh`
#> #          and in `$IRACE_HOME/inst/examples/irace2pyimp/002-TemplateDesign/run.sh`
#> #
#> # irace2pyimp can also be called as an R function instead.  To see usage of the
#> # function, run in R console: 
#> #    > library(irace)
#> #    > ?irace2pyimp 
#> #
#> # The generated files include:
#> #  * `params.pcs` : a text file containing the parameter space definition.
#> #  * `runhistory.json` : a JSON file containing the list of algorithm configurations
#> #    evaluated during the tuning and the performance data obtained.
#> #  * `traj_aclib2.json` : a JSON file containing the best configurations after
#> #     each iteration of irace.
#> #  * `scenario.txt` : a text file containing the definition of the tuning scenario.
#> #  * `instances.txt` : a text file containing the list of instances.
#> #  * `features.csv` : a text file containing instance features.
#> #------------------------------------------------------------------------------
#> 
#> -h --help                Show this help.
#> -n --normalise           Normalise the cost metric values into the range of
#>                          [0,1] before converting to PyImp format. Values:
#>                          {none(default), instance, feature}, which correspond
#>                          to no-normalisation, normalisation based on instance,
#>                          and normalisation based on features, respectively.
#> -o --out-dir             Directory where all generated data are stored.
#>                          Default: ./pyimp-input
#> -f --instance-feature-file  A .csv file containing instance features (one line
#>                          per instance, sorted in the same order as the list of
#>                          instances input to irace). The first line contains
#>                          feature names. By default, instance index is used as a
#>                          feature.
#> -c --filter-conditions   Only extract configurations that satisfies the given
#>                          conditions. The conditions are in R expression format.
#>                          Default: no filter
#> -i --default-configuration-index  Index of default configuration (starting from
#>                          1), used by ablation analysis. Default value: 1
#> -r --irace-data-file     The log .Rdata file generated by irace
#> -ignore --ignore-unsupported  Forbidden configurations and repairConfiguration
#>                          are not supported by the script. Set this flag to 1 to
#>                          ignore them and proceed with your own risk. This may
#>                          cause some unwanted behaviours, e.g., forbidden
#>                          configurations may appear in ablation analysis's path.
#>                          Values: {0,1}