Reads from a file the scenario settings to be used by irace.

The scenario argument is an initial scenario that is overwritten for every setting specified in the file to be read.

Usage

readScenario(filename = "", scenario = list(), params_def = .irace.params.def)

Arguments

filename

character(1)
Filename from which the scenario will be read. If empty, the default scenarioFile is used. An example scenario file is provided in system.file(package="irace", "templates/scenario.txt.tmpl").

scenario

list()
Data structure containing irace settings. The data structure has to be the one returned by the function defaultScenario() or readScenario().

params_def

data.frame()
Definition of the options accepted by the scenario. This should only be modified by packages that wish to extend irace.

Value

The scenario list read from the file. The scenario settings not present in the file are not present in the list, i.e., they are NULL.

See also

Author

Manuel López-Ibáñez and Jérémie Dubois-Lacoste