Read the log file produced by irace (irace.Rdata).

Read the log file produced by irace (irace.Rdata).

Usage

read_logfile(filename, name = "iraceResults")

Arguments

filename

Filename that contains the log file saved by irace. Example: irace.Rdata.

name

Optional argument that allows overriding the default name of the object in the file.

Value

(list())

Examples

irace_results <- read_logfile(system.file("exdata/irace-acotsp.Rdata", package="irace",
                                          mustWork=TRUE))
str(irace_results)
#> List of 11
#>  $ scenario         :List of 58
#>   ..$ scenarioFile             : chr "/home/manu/work/irace/git/devel-examples/vignette-example/scenario.txt"
#>   ..$ configurationsFile       : chr "/home/manu/work/irace/git/devel-examples/vignette-example/default.txt"
#>   ..$ logFile                  : chr "./irace-acotsp.Rdata"
#>   ..$ trainInstancesDir        : chr "./instances"
#>   ..$ trainInstancesFile       : chr "./instances.txt"
#>   ..$ testInstancesDir         : chr "./instances"
#>   ..$ testInstancesFile        : chr "./instances-test.txt"
#>   ..$ testNbElites             : int 5
#>   ..$ testIterationElites      : logi TRUE
#>   ..$ maxExperiments           : int 1000
#>   ..$ seed                     : int 687542627
#>   ..$ execDir                  : chr "./"
#>   ..$ parameterFile            : chr "./parameters.txt"
#>   ..$ recoveryFile             : chr ""
#>   ..$ instances                : chr [1:10] "./instances/1000-1.tsp" "./instances/1000-2.tsp" "./instances/1000-3.tsp" "./instances/1000-4.tsp" ...
#>   ..$ sampleInstances          : logi TRUE
#>   ..$ testInstances            : chr [1:10] "./instances/1000-10.tsp" "./instances/1000-11.tsp" "./instances/1000-12.tsp" "./instances/1000-13.tsp" ...
#>   ..$ testType                 : chr "friedman"
#>   ..$ firstTest                : int 5
#>   ..$ blockSize                : int 1
#>   ..$ eachTest                 : int 1
#>   ..$ targetRunner             : chr "./target-runner"
#>   ..$ targetRunnerLauncher     : chr ""
#>   ..$ targetCmdline            : chr "{configurationID} {instanceID} {seed} {instance} {bound} {targetRunnerArgs}"
#>   ..$ targetRunnerRetries      : int 0
#>   ..$ targetRunnerTimeout      : int 0
#>   ..$ targetRunnerData         : chr ""
#>   ..$ deterministic            : logi FALSE
#>   ..$ minExperiments           : chr NA
#>   ..$ maxTime                  : int 0
#>   ..$ budgetEstimation         : num 0.05
#>   ..$ minMeasurableTime        : num 0.01
#>   ..$ parallel                 : int 2
#>   ..$ loadBalancing            : logi TRUE
#>   ..$ mpi                      : logi FALSE
#>   ..$ batchmode                : chr "0"
#>   ..$ digits                   : int 4
#>   ..$ quiet                    : logi FALSE
#>   ..$ debugLevel               : int 0
#>   ..$ softRestart              : logi TRUE
#>   ..$ softRestartThreshold     : num 1e-04
#>   ..$ elitist                  : logi TRUE
#>   ..$ elitistNewInstances      : int 1
#>   ..$ elitistLimit             : int 2
#>   ..$ capping                  : logi FALSE
#>   ..$ cappingType              : chr "median"
#>   ..$ boundType                : chr "candidate"
#>   ..$ boundDigits              : int 0
#>   ..$ boundPar                 : int 1
#>   ..$ boundAsTimeout           : logi TRUE
#>   ..$ postselection            : num 0
#>   ..$ aclib                    : logi FALSE
#>   ..$ nbIterations             : int 0
#>   ..$ nbExperimentsPerIteration: int 0
#>   ..$ minNbSurvival            : int 0
#>   ..$ nbConfigurations         : int 0
#>   ..$ mu                       : int 5
#>   ..$ confidence               : num 0.95
#>  $ irace.version    : chr "3.5.1.9000.036e23af-dirty"
#>  $ parameters       :List of 15
#>   ..$ names       : chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ types       : Named chr [1:11] "c" "c" "r" "r" ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ switches    : Named chr [1:11] "--" "--localsearch " "--alpha " "--beta " ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ domain      :List of 11
#>   .. ..$ algorithm  : chr [1:5] "as" "mmas" "eas" "ras" ...
#>   .. ..$ localsearch: chr [1:4] "0" "1" "2" "3"
#>   .. ..$ alpha      : num [1:2] 0 5
#>   .. ..$ beta       : num [1:2] 0 10
#>   .. ..$ rho        : num [1:2] 0.01 1
#>   .. ..$ ants       : num [1:2] 5 100
#>   .. ..$ nnls       : num [1:2] 5 50
#>   .. ..$ q0         : num [1:2] 0 1
#>   .. ..$ dlb        : chr [1:2] "0" "1"
#>   .. ..$ rasrank    : num [1:2] 1 100
#>   .. ..$ elitistants: num [1:2] 1 750
#>   ..$ conditions  :List of 11
#>   .. ..$ algorithm  : logi TRUE
#>   .. ..$ localsearch: logi TRUE
#>   .. ..$ alpha      : logi TRUE
#>   .. ..$ beta       : logi TRUE
#>   .. ..$ rho        : logi TRUE
#>   .. ..$ ants       : logi TRUE
#>   .. ..$ nnls       :  expression(localsearch %in% c(1, 2, 3))
#>   .. ..$ q0         :  expression(algorithm %in% c("acs"))
#>   .. ..$ dlb        :  expression(localsearch %in% c(1, 2, 3))
#>   .. ..$ rasrank    :  expression(algorithm %in% c("ras"))
#>   .. ..$ elitistants:  expression(algorithm %in% c("eas"))
#>   ..$ isFixed     : Named logi [1:11] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ transform   :List of 11
#>   .. ..$ algorithm  : chr ""
#>   .. ..$ localsearch: chr ""
#>   .. ..$ alpha      : chr ""
#>   .. ..$ beta       : chr ""
#>   .. ..$ rho        : chr ""
#>   .. ..$ ants       : chr "log"
#>   .. .. ..- attr(*, "lower")= num 1.61
#>   .. .. ..- attr(*, "upper")= num 4.62
#>   .. ..$ nnls       : chr ""
#>   .. ..$ q0         : chr ""
#>   .. ..$ dlb        : chr ""
#>   .. ..$ rasrank    : chr ""
#>   .. ..$ elitistants: chr ""
#>   ..$ isDependent : Named logi [1:11] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ digits      : Named int [1:4] 4 4 4 4
#>   .. ..- attr(*, "names")= chr [1:4] "alpha" "beta" "rho" "q0"
#>   ..$ forbidden   :List of 1
#>   .. ..$ :<bytecode> 
#>   .. .. ..- attr(*, "source")= chr "(alpha == 0) & (beta == 0)"
#>   ..$ depends     :List of 11
#>   .. ..$ algorithm  : chr(0) 
#>   .. ..$ localsearch: chr(0) 
#>   .. ..$ alpha      : chr(0) 
#>   .. ..$ beta       : chr(0) 
#>   .. ..$ rho        : chr(0) 
#>   .. ..$ ants       : chr(0) 
#>   .. ..$ nnls       : chr "localsearch"
#>   .. ..$ q0         : chr "algorithm"
#>   .. ..$ dlb        : chr "localsearch"
#>   .. ..$ rasrank    : chr "algorithm"
#>   .. ..$ elitistants: chr "algorithm"
#>   ..$ hierarchy   : Named num [1:11] 1 1 1 1 1 1 2 2 2 2 ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ nbParameters: int 11
#>   ..$ nbFixed     : int 0
#>   ..$ nbVariable  : int 11
#>  $ iterationElites  : int [1:7] 2 50 85 100 128 153 153
#>  $ allElites        :List of 7
#>   ..$ : int [1:4] 2 29 3 20
#>   ..$ : int [1:5] 50 29 2 47 3
#>   ..$ : int [1:3] 85 70 50
#>   ..$ : int [1:5] 100 85 106 99 112
#>   ..$ : int [1:5] 128 141 133 125 129
#>   ..$ : int [1:5] 153 141 128 133 125
#>   ..$ : int [1:5] 153 141 159 128 125
#>  $ experiments      : num [1:16, 1:159] 28407696 28402347 29166659 28417102 28823388 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:16] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:159] "1" "2" "3" "4" ...
#>  $ experimentLog    : int [1:991, 1:5] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "iteration" "instance" "configuration" "time" ...
#>  $ state            :List of 18
#>   ..$ .Random.seed        : int [1:626] 10403 82 74450096 -488718086 1971160917 1271940195 -1466109272 593734952 -700000720 -249115685 ...
#>   ..$ .irace              :<environment: 0x561a6ea539a0> 
#>   ..$ currentBudget       : num 53
#>   ..$ debugLevel          : int 0
#>   ..$ eliteConfigurations :'data.frame':	5 obs. of  16 variables:
#>   .. ..$ .ID.       : int [1:5] 153 141 159 128 125
#>   .. ..$ algorithm  : chr [1:5] "acs" "acs" "acs" "acs" ...
#>   .. ..$ localsearch: chr [1:5] "3" "3" "3" "3" ...
#>   .. ..$ alpha      : num [1:5] 1.77 1.34 1.38 1.87 2.38
#>   .. ..$ beta       : num [1:5] 6.05 7.49 7.02 7.37 6.78
#>   .. ..$ rho        : num [1:5] 0.172 0.357 0.437 0.409 0.181
#>   .. ..$ ants       : num [1:5] 21 7 19 17 22
#>   .. ..$ nnls       : num [1:5] 19 21 17 12 37
#>   .. ..$ q0         : num [1:5] 0.46 0.1651 0.1531 0.2948 0.0721
#>   .. ..$ dlb        : chr [1:5] "1" "1" "1" "1" ...
#>   .. ..$ rasrank    : num [1:5] NA NA NA NA NA
#>   .. ..$ elitistants: num [1:5] NA NA NA NA NA
#>   .. ..$ .PARENT.   : int [1:5] 128 100 128 100 85
#>   .. ..$ .ALIVE.    : logi [1:5] TRUE TRUE TRUE TRUE TRUE
#>   .. ..$ .RANK.     : num [1:5] 46 53 53 60 62
#>   .. ..$ .WEIGHT.   : num [1:5] 0.3333 0.2667 0.2 0.1333 0.0667
#>   ..$ experimentsUsedSoFar: int 991
#>   ..$ indexIteration      : int 8
#>   ..$ minSurvival         : num 5
#>   ..$ model               :List of 11
#>   .. ..$ algorithm  :List of 5
#>   .. .. ..$ 153: num [1:5] 0.000257 0.000257 0.000257 0.001651 0.997578
#>   .. .. ..$ 141: num [1:5] 0.000257 0.000257 0.000257 0.001651 0.997578
#>   .. .. ..$ 128: num [1:5] 0.000257 0.000257 0.000257 0.001651 0.997578
#>   .. .. ..$ 133: num [1:5] 0.000272 0.000272 0.000272 0.000272 0.998913
#>   .. .. ..$ 125: num [1:5] 0.000257 0.000257 0.000257 0.001651 0.997578
#>   .. ..$ localsearch:List of 5
#>   .. .. ..$ 153: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 141: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 128: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 133: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 125: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. ..$ alpha      :List of 5
#>   .. .. ..$ 153: num [1:2] 0.552 1.773
#>   .. .. ..$ 141: num [1:2] 0.552 1.338
#>   .. .. ..$ 128: num [1:2] 0.552 1.873
#>   .. .. ..$ 133: num [1:2] 0.552 2.323
#>   .. .. ..$ 125: num [1:2] 0.552 2.382
#>   .. ..$ beta       :List of 5
#>   .. .. ..$ 153: num [1:2] 1.1 6.05
#>   .. .. ..$ 141: num [1:2] 1.1 7.49
#>   .. .. ..$ 128: num [1:2] 1.1 7.37
#>   .. .. ..$ 133: num [1:2] 1.1 8.2
#>   .. .. ..$ 125: num [1:2] 1.1 6.78
#>   .. ..$ rho        :List of 5
#>   .. .. ..$ 153: num [1:2] 0.109 0.172
#>   .. .. ..$ 141: num [1:2] 0.109 0.357
#>   .. .. ..$ 128: num [1:2] 0.109 0.409
#>   .. .. ..$ 133: num [1:2] 0.109 0.836
#>   .. .. ..$ 125: num [1:2] 0.109 0.181
#>   .. ..$ ants       :List of 5
#>   .. .. ..$ 153: num [1:2] 0.11 21
#>   .. .. ..$ 141: num [1:2] 0.11 7
#>   .. .. ..$ 128: num [1:2] 0.11 17
#>   .. .. ..$ 133: num [1:2] 0.11 51
#>   .. .. ..$ 125: num [1:2] 0.11 22
#>   .. ..$ nnls       :List of 5
#>   .. .. ..$ 153: num [1:2] 4.97 19
#>   .. .. ..$ 141: num [1:2] 4.97 21
#>   .. .. ..$ 128: num [1:2] 4.97 12
#>   .. .. ..$ 133: num [1:2] 4.97 31
#>   .. .. ..$ 125: num [1:2] 4.97 37
#>   .. ..$ q0         :List of 5
#>   .. .. ..$ 153: num [1:2] 0.11 0.46
#>   .. .. ..$ 141: num [1:2] 0.11 0.165
#>   .. .. ..$ 128: num [1:2] 0.11 0.295
#>   .. .. ..$ 133: num [1:2] 0.11 0.661
#>   .. .. ..$ 125: num [1:2] 0.1105 0.0721
#>   .. ..$ dlb        :List of 5
#>   .. .. ..$ 153: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 141: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 128: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 133: num [1:2] 0.00552 0.99448
#>   .. .. ..$ 125: num [1:2] 0.00102 0.99898
#>   .. ..$ rasrank    :List of 5
#>   .. .. ..$ 153: num [1:2] 10.9 12
#>   .. .. ..$ 141: num [1:2] 10.9 12
#>   .. .. ..$ 128: num [1:2] 10.9 12
#>   .. .. ..$ 133: num [1:2] 10.9 NA
#>   .. .. ..$ 125: num [1:2] 10.9 12
#>   .. ..$ elitistants:List of 5
#>   .. .. ..$ 153: num [1:2] 82.8 NA
#>   .. .. ..$ 141: num [1:2] 82.8 NA
#>   .. .. ..$ 128: num [1:2] 82.8 NA
#>   .. .. ..$ 133: num [1:2] 82.8 NA
#>   .. .. ..$ 125: num [1:2] 82.8 NA
#>   ..$ nbConfigurations    : num 8
#>   ..$ nbIterations        : int 7
#>   ..$ remainingBudget     : int 9
#>   ..$ timeUsed            : num 0
#>   ..$ boundEstimate       : logi NA
#>   ..$ rejectedIDs         : NULL
#>   ..$ completed           : chr "Not enough budget to race more than the minimum configurations"
#>   ..$ elapsed             : Named num [1:3] 9984.1 60.6 5299.4
#>   .. ..- attr(*, "names")= chr [1:3] "user" "system" "wallclock"
#>   ..$ sessionInfo         :List of 13
#>   .. ..$ R.version      :List of 14
#>   .. .. ..$ platform      : chr "x86_64-pc-linux-gnu"
#>   .. .. ..$ arch          : chr "x86_64"
#>   .. .. ..$ os            : chr "linux-gnu"
#>   .. .. ..$ system        : chr "x86_64, linux-gnu"
#>   .. .. ..$ status        : chr ""
#>   .. .. ..$ major         : chr "4"
#>   .. .. ..$ minor         : chr "1.2"
#>   .. .. ..$ year          : chr "2021"
#>   .. .. ..$ month         : chr "11"
#>   .. .. ..$ day           : chr "01"
#>   .. .. ..$ svn rev       : chr "81115"
#>   .. .. ..$ language      : chr "R"
#>   .. .. ..$ version.string: chr "R version 4.1.2 (2021-11-01)"
#>   .. .. ..$ nickname      : chr "Bird Hippie"
#>   .. ..$ platform       : chr "x86_64-pc-linux-gnu (64-bit)"
#>   .. ..$ locale         : chr "LC_CTYPE=en_GB.UTF-8;LC_NUMERIC=C;LC_TIME=en_GB.UTF-8;LC_COLLATE=en_GB.UTF-8;LC_MONETARY=en_GB.UTF-8;LC_MESSAGE"| __truncated__
#>   .. ..$ running        : chr "Ubuntu 22.04.4 LTS"
#>   .. ..$ RNGkind        : chr [1:3] "Mersenne-Twister" "Inversion" "Rejection"
#>   .. ..$ basePkgs       : chr [1:7] "stats" "graphics" "grDevices" "utils" ...
#>   .. ..$ otherPkgs      :List of 1
#>   .. .. ..$ irace:List of 24
#>   .. .. .. ..$ Package                : chr "irace"
#>   .. .. .. ..$ Type                   : chr "Package"
#>   .. .. .. ..$ Title                  : chr "Iterated Racing for Automatic Algorithm Configuration"
#>   .. .. .. ..$ Description            : chr "Iterated race is an extension of the Iterated F-race method for\n             the automatic configuration of op"| __truncated__
#>   .. .. .. ..$ Version                : chr "3.5.1.9000"
#>   .. .. .. ..$ Authors@R              : chr "c(person(\"Manuel\", \"López-Ibáñez\", role = c(\"aut\", \"cre\"),\n                              email = \"man"| __truncated__
#>   .. .. .. ..$ Depends                : chr "R (>= 3.6.0)"
#>   .. .. .. ..$ Imports                : chr "stats, utils, compiler, fs, matrixStats, R6, withr"
#>   .. .. .. ..$ Suggests               : chr "Rmpi (>= 0.6.0), parallel, knitr, testthat (>= 3.1.8),\nParamHelpers, devtools"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ License                : chr "GPL (>= 2)"
#>   .. .. .. ..$ URL                    : chr "https://mlopez-ibanez.github.io/irace/,\nhttps://github.com/MLopez-Ibanez/irace"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/MLopez-Ibanez/irace/issues"
#>   .. .. .. ..$ ByteCompile            : chr "yes"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ Roxygen                : chr "list(markdown = TRUE)"
#>   .. .. .. ..$ RoxygenNote            : chr "7.3.0"
#>   .. .. .. ..$ SystemRequirements     : chr "GNU make"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2024-03-22 14:05:24 UTC; manu"
#>   .. .. .. ..$ Author                 : chr "Manuel López-Ibáñez [aut, cre]\n    (<https://orcid.org/0000-0001-9974-1295>),\n  Jérémie Dubois-Lacoste [aut],"| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Manuel López-Ibáñez <manuel.lopez-ibanez@manchester.ac.uk>"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2024-03-22 14:05:25 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/irace/Meta/package.rds"
#>   .. ..$ loadedOnly     :List of 17
#>   .. .. ..$ matrixStats:List of 27
#>   .. .. .. ..$ Package          : chr "matrixStats"
#>   .. .. .. ..$ Version          : chr "1.2.0"
#>   .. .. .. ..$ Depends          : chr "R (>= 2.12.0)"
#>   .. .. .. ..$ Suggests         : chr "base64enc, ggplot2, knitr, markdown, microbenchmark,\nR.devices, R.rsp"
#>   .. .. .. ..$ VignetteBuilder  : chr "R.rsp"
#>   .. .. .. ..$ Title            : chr "Functions that Apply to Rows and Columns of Matrices (and to\nVectors)"
#>   .. .. .. ..$ Authors@R        : chr "c(\n  person(\"Henrik\", \"Bengtsson\", role=c(\"aut\", \"cre\", \"cph\"),\n                                 em"| __truncated__
#>   .. .. .. ..$ Author           : chr "Henrik Bengtsson [aut, cre, cph],\n  Constantin Ahlmann-Eltze [ctb],\n  Hector Corrada Bravo [ctb],\n  Robert G"| __truncated__
#>   .. .. .. ..$ Maintainer       : chr "Henrik Bengtsson <henrikb@braju.com>"
#>   .. .. .. ..$ Description      : chr "High-performing functions operating on rows and columns of matrices, e.g. col / rowMedians(), col / rowRanks(),"| __truncated__
#>   .. .. .. ..$ License          : chr "Artistic-2.0"
#>   .. .. .. ..$ LazyLoad         : chr "TRUE"
#>   .. .. .. ..$ NeedsCompilation : chr "yes"
#>   .. .. .. ..$ ByteCompile      : chr "TRUE"
#>   .. .. .. ..$ URL              : chr "https://github.com/HenrikBengtsson/matrixStats"
#>   .. .. .. ..$ BugReports       : chr "https://github.com/HenrikBengtsson/matrixStats/issues"
#>   .. .. .. ..$ RoxygenNote      : chr "7.2.3"
#>   .. .. .. ..$ Packaged         : chr "2023-12-11 19:55:37 UTC; henrik"
#>   .. .. .. ..$ Repository       : chr "CRAN"
#>   .. .. .. ..$ Date/Publication : chr "2023-12-11 22:30:09 UTC"
#>   .. .. .. ..$ Built            : chr "R 4.1.2; x86_64-pc-linux-gnu; 2024-01-23 14:02:33 UTC; unix"
#>   .. .. .. ..$ RemoteType       : chr "standard"
#>   .. .. .. ..$ RemotePkgRef     : chr "matrixStats"
#>   .. .. .. ..$ RemoteRef        : chr "matrixStats"
#>   .. .. .. ..$ RemoteRepos      : chr "https://cloud.r-project.org"
#>   .. .. .. ..$ RemotePkgPlatform: chr "source"
#>   .. .. .. ..$ RemoteSha        : chr "1.2.0"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/matrixStats/Meta/package.rds"
#>   .. .. ..$ fansi      :List of 22
#>   .. .. .. ..$ Package         : chr "fansi"
#>   .. .. .. ..$ Title           : chr "ANSI Control Sequence Aware String Functions"
#>   .. .. .. ..$ Description     : chr "Counterparts to R string manipulation functions that account for\n   the effects of ANSI text formatting control sequences."
#>   .. .. .. ..$ Version         : chr "1.0.6"
#>   .. .. .. ..$ Authors@R       : chr "c(\n    person(\"Brodie\", \"Gaslam\", email=\"brodie.gaslam@yahoo.com\",\n    role=c(\"aut\", \"cre\")),\n    "| __truncated__
#>   .. .. .. ..$ Depends         : chr "R (>= 3.1.0)"
#>   .. .. .. ..$ License         : chr "GPL-2 | GPL-3"
#>   .. .. .. ..$ URL             : chr "https://github.com/brodieG/fansi"
#>   .. .. .. ..$ BugReports      : chr "https://github.com/brodieG/fansi/issues"
#>   .. .. .. ..$ VignetteBuilder : chr "knitr"
#>   .. .. .. ..$ Suggests        : chr "unitizer, knitr, rmarkdown"
#>   .. .. .. ..$ Imports         : chr "grDevices, utils"
#>   .. .. .. ..$ RoxygenNote     : chr "7.2.3"
#>   .. .. .. ..$ Encoding        : chr "UTF-8"
#>   .. .. .. ..$ Collate         : chr "'constants.R' 'fansi-package.R' 'internal.R' 'load.R' 'misc.R'\n'nchar.R' 'strwrap.R' 'strtrim.R' 'strsplit.R' "| __truncated__
#>   .. .. .. ..$ NeedsCompilation: chr "yes"
#>   .. .. .. ..$ Packaged        : chr "2023-12-06 00:59:41 UTC; bg"
#>   .. .. .. ..$ Author          : chr "Brodie Gaslam [aut, cre],\n  Elliott Sales De Andrade [ctb],\n  R Core Team [cph] (UTF8 byte length calcs from src/util.c)"
#>   .. .. .. ..$ Maintainer      : chr "Brodie Gaslam <brodie.gaslam@yahoo.com>"
#>   .. .. .. ..$ Repository      : chr "CRAN"
#>   .. .. .. ..$ Date/Publication: chr "2023-12-08 03:30:02 UTC"
#>   .. .. .. ..$ Built           : chr "R 4.1.2; x86_64-pc-linux-gnu; 2024-03-20 16:56:02 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/fansi/Meta/package.rds"
#>   .. .. ..$ utf8       :List of 21
#>   .. .. .. ..$ Package                : chr "utf8"
#>   .. .. .. ..$ Title                  : chr "Unicode Text Processing"
#>   .. .. .. ..$ Version                : chr "1.2.4"
#>   .. .. .. ..$ Authors@R              : chr "\n    c(person(given = c(\"Patrick\", \"O.\"),\n             family = \"Perry\",\n             role = c(\"aut\""| __truncated__
#>   .. .. .. ..$ Description            : chr "Process and print 'UTF-8' encoded international\n    text (Unicode). Input, validate, normalize, encode, format"| __truncated__
#>   .. .. .. ..$ License                : chr "Apache License (== 2.0) | file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://ptrckprry.com/r-utf8/, https://github.com/patperry/r-utf8"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/patperry/r-utf8/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 2.10)"
#>   .. .. .. ..$ Suggests               : chr "cli, covr, knitr, rlang, rmarkdown, testthat (>= 3.0.0),\nwithr"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr, rmarkdown"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2023-10-22 13:43:19 UTC; kirill"
#>   .. .. .. ..$ Author                 : chr "Patrick O. Perry [aut, cph],\n  Kirill Müller [cre],\n  Unicode, Inc. [cph, dtc] (Unicode Character Database)"
#>   .. .. .. ..$ Maintainer             : chr "Kirill Müller <kirill@cynkra.com>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2023-10-22 21:50:02 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2024-03-20 16:55:57 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/utf8/Meta/package.rds"
#>   .. .. ..$ withr      :List of 25
#>   .. .. .. ..$ Package                : chr "withr"
#>   .. .. .. ..$ Title                  : chr "Run Code 'With' Temporarily Modified Global State"
#>   .. .. .. ..$ Version                : chr "2.99.0.9000"
#>   .. .. .. ..$ Authors@R              : chr "\n    c(person(given = \"Jim\",\n             family = \"Hester\",\n             role = \"aut\"),\n      person"| __truncated__
#>   .. .. .. ..$ Description            : chr "A set of functions to run code 'with' safely and\n    temporarily modified global state. Many of these function"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://withr.r-lib.org, https://github.com/r-lib/withr#readme"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/withr/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.5.0)"
#>   .. .. .. ..$ Imports                : chr "graphics, grDevices,"
#>   .. .. .. ..$ Suggests               : chr "callr, covr, DBI, knitr, lattice, methods, rlang, rmarkdown\n(>= 2.12), RSQLite, testthat (>= 3.0.0)"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ Roxygen                : chr "list(markdown = TRUE)"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ Collate                : chr "'aaa.R' 'collate.R' 'connection.R' 'db.R' 'defer-exit.R'\n'standalone-defer.R' 'defer.R' 'wrap.R' 'local_.R' 'w"| __truncated__
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Config/Needs/website   : chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ RemoteType             : chr "local"
#>   .. .. .. ..$ RemoteUrl              : chr "/home/manu/work/src/withr"
#>   .. .. .. ..$ NeedsCompilation       : chr "no"
#>   .. .. .. ..$ Packaged               : chr "2023-12-19 21:18:36 UTC; manu"
#>   .. .. .. ..$ Author                 : chr "Jim Hester [aut],\n  Lionel Henry [aut, cre],\n  Kirill Müller [aut],\n  Kevin Ushey [aut],\n  Hadley Wickham ["| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Lionel Henry <lionel@posit.co>"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; ; 2023-12-19 21:18:37 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/withr/Meta/package.rds"
#>   .. .. ..$ R6         :List of 18
#>   .. .. .. ..$ Package         : chr "R6"
#>   .. .. .. ..$ Title           : chr "Encapsulated Classes with Reference Semantics"
#>   .. .. .. ..$ Version         : chr "2.5.1"
#>   .. .. .. ..$ Authors@R       : chr "person(\"Winston\", \"Chang\", role = c(\"aut\", \"cre\"), email = \"winston@stdout.org\")"
#>   .. .. .. ..$ Description     : chr "Creates classes with reference semantics, similar to R's built-in\n    reference classes. Compared to reference"| __truncated__
#>   .. .. .. ..$ Depends         : chr "R (>= 3.0)"
#>   .. .. .. ..$ Suggests        : chr "testthat, pryr"
#>   .. .. .. ..$ License         : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL             : chr "https://r6.r-lib.org, https://github.com/r-lib/R6/"
#>   .. .. .. ..$ BugReports      : chr "https://github.com/r-lib/R6/issues"
#>   .. .. .. ..$ RoxygenNote     : chr "7.1.1"
#>   .. .. .. ..$ NeedsCompilation: chr "no"
#>   .. .. .. ..$ Packaged        : chr "2021-08-06 20:18:46 UTC; winston"
#>   .. .. .. ..$ Author          : chr "Winston Chang [aut, cre]"
#>   .. .. .. ..$ Maintainer      : chr "Winston Chang <winston@stdout.org>"
#>   .. .. .. ..$ Repository      : chr "CRAN"
#>   .. .. .. ..$ Date/Publication: chr "2021-08-19 14:00:05 UTC"
#>   .. .. .. ..$ Built           : chr "R 4.1.1; ; 'Sat, 21 Aug 2021 18:04:39 +0000'; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/site-library/R6/Meta/package.rds"
#>   .. .. ..$ lifecycle  :List of 23
#>   .. .. .. ..$ Package                : chr "lifecycle"
#>   .. .. .. ..$ Title                  : chr "Manage the Life Cycle of your Package Functions"
#>   .. .. .. ..$ Version                : chr "1.0.4"
#>   .. .. .. ..$ Authors@R              : chr "c(\n    person(\"Lionel\", \"Henry\", , \"lionel@posit.co\", role = c(\"aut\", \"cre\")),\n    person(\"Hadley\"| __truncated__
#>   .. .. .. ..$ Description            : chr "Manage the life cycle of your exported functions with shared\n    conventions, documentation badges, and user-f"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://lifecycle.r-lib.org/, https://github.com/r-lib/lifecycle"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/lifecycle/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.6)"
#>   .. .. .. ..$ Imports                : chr "cli (>= 3.4.0), glue, rlang (>= 1.1.0)"
#>   .. .. .. ..$ Suggests               : chr "covr, crayon, knitr, lintr, rmarkdown, testthat (>= 3.0.1),\ntibble, tidyverse, tools, vctrs, withr"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ Config/Needs/website   : chr "tidyverse/tidytemplate, usethis"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.1"
#>   .. .. .. ..$ NeedsCompilation       : chr "no"
#>   .. .. .. ..$ Packaged               : chr "2023-11-06 16:07:36 UTC; lionel"
#>   .. .. .. ..$ Author                 : chr "Lionel Henry [aut, cre],\n  Hadley Wickham [aut] (<https://orcid.org/0000-0003-4757-117X>),\n  Posit Software, PBC [cph, fnd]"
#>   .. .. .. ..$ Maintainer             : chr "Lionel Henry <lionel@posit.co>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2023-11-07 10:10:10 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; ; 2023-12-03 12:21:15 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/lifecycle/Meta/package.rds"
#>   .. .. ..$ magrittr   :List of 23
#>   .. .. .. ..$ Type                : chr "Package"
#>   .. .. .. ..$ Package             : chr "magrittr"
#>   .. .. .. ..$ Title               : chr "A Forward-Pipe Operator for R"
#>   .. .. .. ..$ Version             : chr "2.0.3"
#>   .. .. .. ..$ Authors@R           : chr "c(\n    person(\"Stefan Milton\", \"Bache\", , \"stefan@stefanbache.dk\", role = c(\"aut\", \"cph\"),\n        "| __truncated__
#>   .. .. .. ..$ Description         : chr "Provides a mechanism for chaining commands with a new\n    forward-pipe operator, %>%. This operator will forwa"| __truncated__
#>   .. .. .. ..$ License             : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                 : chr "https://magrittr.tidyverse.org,\nhttps://github.com/tidyverse/magrittr"
#>   .. .. .. ..$ BugReports          : chr "https://github.com/tidyverse/magrittr/issues"
#>   .. .. .. ..$ Depends             : chr "R (>= 3.4.0)"
#>   .. .. .. ..$ Suggests            : chr "covr, knitr, rlang, rmarkdown, testthat"
#>   .. .. .. ..$ VignetteBuilder     : chr "knitr"
#>   .. .. .. ..$ ByteCompile         : chr "Yes"
#>   .. .. .. ..$ Config/Needs/website: chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ Encoding            : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote         : chr "7.1.2"
#>   .. .. .. ..$ NeedsCompilation    : chr "yes"
#>   .. .. .. ..$ Packaged            : chr "2022-03-29 09:34:37 UTC; lionel"
#>   .. .. .. ..$ Author              : chr "Stefan Milton Bache [aut, cph] (Original author and creator of\n    magrittr),\n  Hadley Wickham [aut],\n  Lion"| __truncated__
#>   .. .. .. ..$ Maintainer          : chr "Lionel Henry <lionel@rstudio.com>"
#>   .. .. .. ..$ Repository          : chr "CRAN"
#>   .. .. .. ..$ Date/Publication    : chr "2022-03-30 07:30:09 UTC"
#>   .. .. .. ..$ Built               : chr "R 4.1.2; x86_64-pc-linux-gnu; 2023-12-03 12:20:13 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/magrittr/Meta/package.rds"
#>   .. .. ..$ pillar     :List of 27
#>   .. .. .. ..$ Package                    : chr "pillar"
#>   .. .. .. ..$ Title                      : chr "Coloured Formatting for Columns"
#>   .. .. .. ..$ Version                    : chr "1.9.0"
#>   .. .. .. ..$ Authors@R                  : chr "\n    c(person(given = \"Kirill\",\n             family = \"M\\u00fcller\",\n             role = c(\"aut\", \"c"| __truncated__
#>   .. .. .. ..$ Description                : chr "Provides 'pillar' and 'colonnade' generics designed\n    for formatting columns of data using the full range of"| __truncated__
#>   .. .. .. ..$ License                    : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                        : chr "https://pillar.r-lib.org/, https://github.com/r-lib/pillar"
#>   .. .. .. ..$ BugReports                 : chr "https://github.com/r-lib/pillar/issues"
#>   .. .. .. ..$ Imports                    : chr "cli (>= 2.3.0), fansi, glue, lifecycle, rlang (>= 1.0.2), utf8\n(>= 1.1.0), utils, vctrs (>= 0.5.0)"
#>   .. .. .. ..$ Suggests                   : chr "bit64, DBI, debugme, DiagrammeR, dplyr, formattable, ggplot2,\nknitr, lubridate, nanotime, nycflights13, palmer"| __truncated__
#>   .. .. .. ..$ VignetteBuilder            : chr "knitr"
#>   .. .. .. ..$ Encoding                   : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote                : chr "7.2.3"
#>   .. .. .. ..$ Config/testthat/edition    : chr "3"
#>   .. .. .. ..$ Config/testthat/parallel   : chr "true"
#>   .. .. .. ..$ Config/testthat/start-first: chr "format_multi_fuzz, format_multi_fuzz_2,\nformat_multi, ctl_colonnade, ctl_colonnade_1, ctl_colonnade_2"
#>   .. .. .. ..$ Config/autostyle/scope     : chr "line_breaks"
#>   .. .. .. ..$ Config/autostyle/strict    : chr "true"
#>   .. .. .. ..$ Config/gha/extra-packages  : chr "DiagrammeR=?ignore-before-r=3.5.0"
#>   .. .. .. ..$ Config/Needs/website       : chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ NeedsCompilation           : chr "no"
#>   .. .. .. ..$ Packaged                   : chr "2023-03-21 08:42:46 UTC; kirill"
#>   .. .. .. ..$ Author                     : chr "Kirill Müller [aut, cre] (<https://orcid.org/0000-0002-1416-3412>),\n  Hadley Wickham [aut],\n  RStudio [cph]"
#>   .. .. .. ..$ Maintainer                 : chr "Kirill Müller <kirill@cynkra.com>"
#>   .. .. .. ..$ Repository                 : chr "CRAN"
#>   .. .. .. ..$ Date/Publication           : chr "2023-03-22 08:10:02 UTC"
#>   .. .. .. ..$ Built                      : chr "R 4.1.2; ; 2023-05-17 15:15:45 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/pillar/Meta/package.rds"
#>   .. .. ..$ rlang      :List of 25
#>   .. .. .. ..$ Package                : chr "rlang"
#>   .. .. .. ..$ Version                : chr "1.1.3"
#>   .. .. .. ..$ Title                  : chr "Functions for Base Types and Core R and 'Tidyverse' Features"
#>   .. .. .. ..$ Description            : chr "A toolbox for working with base types, core R features\n  like the condition system, and core 'Tidyverse' featu"| __truncated__
#>   .. .. .. ..$ Authors@R              : chr "c(\n    person(\"Lionel\", \"Henry\", ,\"lionel@posit.co\", c(\"aut\", \"cre\")),\n    person(\"Hadley\", \"Wic"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ ByteCompile            : chr "true"
#>   .. .. .. ..$ Biarch                 : chr "true"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.5.0)"
#>   .. .. .. ..$ Imports                : chr "utils"
#>   .. .. .. ..$ Suggests               : chr "cli (>= 3.1.0), covr, crayon, fs, glue, knitr, magrittr,\nmethods, pillar, rmarkdown, stats, testthat (>= 3.0.0"| __truncated__
#>   .. .. .. ..$ Enhances               : chr "winch"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ URL                    : chr "https://rlang.r-lib.org, https://github.com/r-lib/rlang"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/rlang/issues"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Config/Needs/website   : chr "dplyr, tidyverse/tidytemplate"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2024-01-09 13:33:10 UTC; lionel"
#>   .. .. .. ..$ Author                 : chr "Lionel Henry [aut, cre],\n  Hadley Wickham [aut],\n  mikefc [cph] (Hash implementation based on Mike's xxhashli"| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Lionel Henry <lionel@posit.co>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2024-01-10 12:00:02 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2024-01-19 17:47:25 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/rlang/Meta/package.rds"
#>   .. .. ..$ cli        :List of 22
#>   .. .. .. ..$ Package                : chr "cli"
#>   .. .. .. ..$ Title                  : chr "Helpers for Developing Command Line Interfaces"
#>   .. .. .. ..$ Version                : chr "3.6.2"
#>   .. .. .. ..$ Authors@R              : chr "c(\n    person(\"Gábor\", \"Csárdi\", , \"csardi.gabor@gmail.com\", role = c(\"aut\", \"cre\")),\n    person(\""| __truncated__
#>   .. .. .. ..$ Description            : chr "A suite of tools to build attractive command line interfaces\n    ('CLIs'), from semantic elements: headings, l"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://cli.r-lib.org, https://github.com/r-lib/cli"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/cli/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.4)"
#>   .. .. .. ..$ Imports                : chr "utils"
#>   .. .. .. ..$ Suggests               : chr "callr, covr, crayon, digest, glue (>= 1.6.0), grDevices,\nhtmltools, htmlwidgets, knitr, methods, mockery, proc"| __truncated__
#>   .. .. .. ..$ Config/Needs/website   : chr "r-lib/asciicast, bench, brio, cpp11, decor, desc,\nfansi, prettyunits, sessioninfo, tidyverse/tidytemplate,\nusethis, vctrs"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2023-12-10 22:38:10 UTC; gaborcsardi"
#>   .. .. .. ..$ Author                 : chr "Gábor Csárdi [aut, cre],\n  Hadley Wickham [ctb],\n  Kirill Müller [ctb],\n  Salim Brüggemann [ctb] (<https://o"| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Gábor Csárdi <csardi.gabor@gmail.com>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2023-12-11 07:40:13 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2023-12-11 09:11:08 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/cli/Meta/package.rds"
#>   .. .. ..$ fs         :List of 27
#>   .. .. .. ..$ Package                : chr "fs"
#>   .. .. .. ..$ Title                  : chr "Cross-Platform File System Operations Based on 'libuv'"
#>   .. .. .. ..$ Version                : chr "1.6.3"
#>   .. .. .. ..$ Authors@R              : chr "c(\n    person(\"Jim\", \"Hester\", role = \"aut\"),\n    person(\"Hadley\", \"Wickham\", , \"hadley@rstudio.co"| __truncated__
#>   .. .. .. ..$ Description            : chr "A cross-platform interface to file system operations, built\n    on top of the 'libuv' C library."
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://fs.r-lib.org, https://github.com/r-lib/fs"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/fs/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.4)"
#>   .. .. .. ..$ Imports                : chr "methods"
#>   .. .. .. ..$ Suggests               : chr "covr, crayon, knitr, pillar (>= 1.0.0), rmarkdown, spelling,\ntestthat (>= 3.0.0), tibble (>= 1.1.0), vctrs (>= 0.3.0), withr"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ ByteCompile            : chr "true"
#>   .. .. .. ..$ Copyright              : chr "file COPYRIGHTS"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ Language               : chr "en-US"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ SystemRequirements     : chr "GNU make"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Config/Needs/website   : chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2023-07-10 16:24:00 UTC; gaborcsardi"
#>   .. .. .. ..$ Author                 : chr "Jim Hester [aut],\n  Hadley Wickham [aut],\n  Gábor Csárdi [aut, cre],\n  libuv project contributors [cph] (lib"| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Gábor Csárdi <csardi.gabor@gmail.com>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2023-07-20 10:30:02 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2023-12-03 12:20:37 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/fs/Meta/package.rds"
#>   .. .. ..$ vctrs      :List of 24
#>   .. .. .. ..$ Package                : chr "vctrs"
#>   .. .. .. ..$ Title                  : chr "Vector Helpers"
#>   .. .. .. ..$ Version                : chr "0.6.5"
#>   .. .. .. ..$ Authors@R              : chr "c(\n    person(\"Hadley\", \"Wickham\", , \"hadley@posit.co\", role = \"aut\"),\n    person(\"Lionel\", \"Henry"| __truncated__
#>   .. .. .. ..$ Description            : chr "Defines new notions of prototype and size that are used to\n    provide tools for consistent and well-founded t"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://vctrs.r-lib.org/, https://github.com/r-lib/vctrs"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/vctrs/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.5.0)"
#>   .. .. .. ..$ Imports                : chr "cli (>= 3.4.0), glue, lifecycle (>= 1.0.3), rlang (>= 1.1.0)"
#>   .. .. .. ..$ Suggests               : chr "bit64, covr, crayon, dplyr (>= 0.8.5), generics, knitr,\npillar (>= 1.4.4), pkgdown (>= 2.0.1), rmarkdown, test"| __truncated__
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ Config/Needs/website   : chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ Language               : chr "en-GB"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2023-12-01 16:27:12 UTC; davis"
#>   .. .. .. ..$ Author                 : chr "Hadley Wickham [aut],\n  Lionel Henry [aut],\n  Davis Vaughan [aut, cre],\n  data.table team [cph] (Radix sort "| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Davis Vaughan <davis@posit.co>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2023-12-01 23:50:02 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2023-12-03 12:21:32 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/vctrs/Meta/package.rds"
#>   .. .. ..$ glue       :List of 24
#>   .. .. .. ..$ Package                : chr "glue"
#>   .. .. .. ..$ Title                  : chr "Interpreted String Literals"
#>   .. .. .. ..$ Version                : chr "1.7.0"
#>   .. .. .. ..$ Authors@R              : chr "c(\n    person(\"Jim\", \"Hester\", role = \"aut\",\n           comment = c(ORCID = \"0000-0002-2739-7082\")),\"| __truncated__
#>   .. .. .. ..$ Description            : chr "An implementation of interpreted string literals, inspired by\n    Python's Literal String Interpolation\n    <"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://glue.tidyverse.org/, https://github.com/tidyverse/glue"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/tidyverse/glue/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.6)"
#>   .. .. .. ..$ Imports                : chr "methods"
#>   .. .. .. ..$ Suggests               : chr "crayon, DBI (>= 1.2.0), dplyr, knitr, magrittr, rlang,\nrmarkdown, RSQLite, testthat (>= 3.2.0), vctrs (>= 0.3."| __truncated__
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ ByteCompile            : chr "true"
#>   .. .. .. ..$ Config/Needs/website   : chr "bench, forcats, ggbeeswarm, ggplot2, R.utils,\nrprintf, tidyr, tidyverse/tidytemplate"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3.9000"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2024-01-08 16:10:57 UTC; jenny"
#>   .. .. .. ..$ Author                 : chr "Jim Hester [aut] (<https://orcid.org/0000-0002-2739-7082>),\n  Jennifer Bryan [aut, cre] (<https://orcid.org/00"| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Jennifer Bryan <jenny@posit.co>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2024-01-09 23:13:08 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2024-01-19 17:47:40 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/glue/Meta/package.rds"
#>   .. .. ..$ parallel   :List of 14
#>   .. .. .. ..$ Package         : chr "parallel"
#>   .. .. .. ..$ Version         : chr "4.1.2"
#>   .. .. .. ..$ Priority        : chr "base"
#>   .. .. .. ..$ Title           : chr "Support for Parallel computation in R"
#>   .. .. .. ..$ Author          : chr "R Core Team"
#>   .. .. .. ..$ Maintainer      : chr "R Core Team <do-use-Contact-address@r-project.org>"
#>   .. .. .. ..$ Contact         : chr "R-help mailing list <r-help@r-project.org>"
#>   .. .. .. ..$ Description     : chr "Support for parallel computation, including by forking\n   (taken from package multicore), by sockets (taken fr"| __truncated__
#>   .. .. .. ..$ License         : chr "Part of R 4.1.2"
#>   .. .. .. ..$ Imports         : chr "tools, compiler"
#>   .. .. .. ..$ Suggests        : chr "methods"
#>   .. .. .. ..$ Enhances        : chr "snow, nws, Rmpi"
#>   .. .. .. ..$ NeedsCompilation: chr "yes"
#>   .. .. .. ..$ Built           : chr "R 4.1.2; x86_64-pc-linux-gnu; 2022-02-09 05:09:20 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/library/parallel/Meta/package.rds"
#>   .. .. ..$ compiler   :List of 10
#>   .. .. .. ..$ Package    : chr "compiler"
#>   .. .. .. ..$ Version    : chr "4.1.2"
#>   .. .. .. ..$ Priority   : chr "base"
#>   .. .. .. ..$ Title      : chr "The R Compiler Package"
#>   .. .. .. ..$ Author     : chr "Luke Tierney <luke-tierney@uiowa.edu>"
#>   .. .. .. ..$ Maintainer : chr "R Core Team <do-use-Contact-address@r-project.org>"
#>   .. .. .. ..$ Contact    : chr "R-help mailing list <r-help@r-project.org>"
#>   .. .. .. ..$ Description: chr "Byte code compiler for R."
#>   .. .. .. ..$ License    : chr "Part of R 4.1.2"
#>   .. .. .. ..$ Built      : chr "R 4.1.2; ; 2022-02-09 05:09:18 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/library/compiler/Meta/package.rds"
#>   .. .. ..$ pkgconfig  :List of 18
#>   .. .. .. ..$ Package         : chr "pkgconfig"
#>   .. .. .. ..$ Title           : chr "Private Configuration for 'R' Packages"
#>   .. .. .. ..$ Version         : chr "2.0.3"
#>   .. .. .. ..$ Author          : chr "Gábor Csárdi"
#>   .. .. .. ..$ Maintainer      : chr "Gábor Csárdi <csardi.gabor@gmail.com>"
#>   .. .. .. ..$ Description     : chr "Set configuration options on a per-package basis.\n    Options set by a given package only apply to that packag"| __truncated__
#>   .. .. .. ..$ License         : chr "MIT + file LICENSE"
#>   .. .. .. ..$ LazyData        : chr "true"
#>   .. .. .. ..$ Imports         : chr "utils"
#>   .. .. .. ..$ Suggests        : chr "covr, testthat, disposables (>= 1.0.3)"
#>   .. .. .. ..$ URL             : chr "https://github.com/r-lib/pkgconfig#readme"
#>   .. .. .. ..$ BugReports      : chr "https://github.com/r-lib/pkgconfig/issues"
#>   .. .. .. ..$ Encoding        : chr "UTF-8"
#>   .. .. .. ..$ NeedsCompilation: chr "no"
#>   .. .. .. ..$ Packaged        : chr "2019-09-22 08:42:40 UTC; gaborcsardi"
#>   .. .. .. ..$ Repository      : chr "CRAN"
#>   .. .. .. ..$ Date/Publication: chr "2019-09-22 09:20:02 UTC"
#>   .. .. .. ..$ Built           : chr "R 4.0.1; ; 'Sat, 30 May 2020 10:19:57 +0000'; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/site-library/pkgconfig/Meta/package.rds"
#>   .. .. ..$ tibble     :List of 28
#>   .. .. .. ..$ Package                    : chr "tibble"
#>   .. .. .. ..$ Title                      : chr "Simple Data Frames"
#>   .. .. .. ..$ Version                    : chr "3.2.1"
#>   .. .. .. ..$ Authors@R                  : chr "\n    c(person(given = \"Kirill\",\n             family = \"M\\u00fcller\",\n             role = c(\"aut\", \"c"| __truncated__
#>   .. .. .. ..$ Description                : chr "Provides a 'tbl_df' class (the 'tibble') with stricter checking and better formatting than the traditional\n    data frame."
#>   .. .. .. ..$ License                    : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                        : chr "https://tibble.tidyverse.org/, https://github.com/tidyverse/tibble"
#>   .. .. .. ..$ BugReports                 : chr "https://github.com/tidyverse/tibble/issues"
#>   .. .. .. ..$ Depends                    : chr "R (>= 3.4.0)"
#>   .. .. .. ..$ Imports                    : chr "fansi (>= 0.4.0), lifecycle (>= 1.0.0), magrittr, methods,\npillar (>= 1.8.1), pkgconfig, rlang (>= 1.0.2), uti"| __truncated__
#>   .. .. .. ..$ Suggests                   : chr "bench, bit64, blob, brio, callr, cli, covr, crayon (>=\n1.3.4), DiagrammeR, dplyr, evaluate, formattable, ggplo"| __truncated__
#>   .. .. .. ..$ VignetteBuilder            : chr "knitr"
#>   .. .. .. ..$ Encoding                   : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote                : chr "7.2.3"
#>   .. .. .. ..$ Config/testthat/edition    : chr "3"
#>   .. .. .. ..$ Config/testthat/parallel   : chr "true"
#>   .. .. .. ..$ Config/testthat/start-first: chr "vignette-formats, as_tibble, add,\ninvariants"
#>   .. .. .. ..$ Config/autostyle/scope     : chr "line_breaks"
#>   .. .. .. ..$ Config/autostyle/strict    : chr "true"
#>   .. .. .. ..$ Config/autostyle/rmd       : chr "false"
#>   .. .. .. ..$ Config/Needs/website       : chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ NeedsCompilation           : chr "yes"
#>   .. .. .. ..$ Packaged                   : chr "2023-03-19 09:23:10 UTC; kirill"
#>   .. .. .. ..$ Author                     : chr "Kirill Müller [aut, cre] (<https://orcid.org/0000-0002-1416-3412>),\n  Hadley Wickham [aut],\n  Romain Francois"| __truncated__
#>   .. .. .. ..$ Maintainer                 : chr "Kirill Müller <kirill@cynkra.com>"
#>   .. .. .. ..$ Repository                 : chr "CRAN"
#>   .. .. .. ..$ Date/Publication           : chr "2023-03-20 06:30:02 UTC"
#>   .. .. .. ..$ Built                      : chr "R 4.1.2; x86_64-pc-linux-gnu; 2023-05-17 15:15:52 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/tibble/Meta/package.rds"
#>   .. ..$ matprod        : chr "default"
#>   .. ..$ BLAS           : chr "/usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3"
#>   .. ..$ LAPACK         : chr "/usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so"
#>   .. ..$ system.codepage: chr "NULL"
#>   .. ..$ codepage       : chr "NULL"
#>   .. ..- attr(*, "class")= chr "sessionInfo"
#>  $ allConfigurations:'data.frame':	159 obs. of  13 variables:
#>   ..$ .ID.       : int [1:159] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ algorithm  : chr [1:159] "as" "acs" "eas" "ras" ...
#>   ..$ localsearch: chr [1:159] "0" "3" "3" "1" ...
#>   ..$ alpha      : num [1:159] 1 1.13 3.72 4.84 4.21 ...
#>   ..$ beta       : num [1:159] 1 3.35 5.52 5.93 5.23 ...
#>   ..$ rho        : num [1:159] 0.95 0.647 0.6 0.761 0.761 ...
#>   ..$ ants       : num [1:159] 10 36 75 5 22 16 40 15 67 7 ...
#>   ..$ nnls       : num [1:159] NA 43 29 40 8 26 8 11 28 27 ...
#>   ..$ q0         : num [1:159] NA 0.905 NA NA NA ...
#>   ..$ dlb        : chr [1:159] NA "0" "1" "0" ...
#>   ..$ rasrank    : num [1:159] NA NA NA 61 49 NA NA NA NA NA ...
#>   ..$ elitistants: num [1:159] NA NA 91 NA NA NA NA 336 350 NA ...
#>   ..$ .PARENT.   : int [1:159] NA NA NA NA NA NA NA NA NA NA ...
#>  $ softRestart      : logi [1:7] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ testing          :List of 2
#>   ..$ experiments: num [1:10, 1:19] 23470396 23236583 23201879 23080399 23329487 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "1t" "2t" "3t" "4t" ...
#>   .. .. ..$ : chr [1:19] "2" "29" "3" "20" ...
#>   ..$ seeds      : Named int [1:10] 1722284619 1499602178 107209901 813566642 1204897224 1411615557 1126399136 672480491 1843150181 1527744095
#>   .. ..- attr(*, "names")= chr [1:10] "1t" "2t" "3t" "4t" ...