Update filesystem paths of a scenario consistently.

This function should be used to change the filesystem paths stored in a scenario object. Useful when moving a scenario from one computer to another.

scenario.update.paths(scenario, from, to, fixed = TRUE)

Arguments

scenario

(list())
Data structure containing irace settings. The data structure has to be the one returned by the function defaultScenario or readScenario.

from

character string containing a regular expression (or character string for fixed = TRUE) to be matched.

to

the replacement string.character string. For fixed = FALSE this can include backreferences "\1" to "\9" to parenthesized subexpressions of from.

fixed

logical. If TRUE, from is a string to be matched as is.

Value

The updated scenario

See also

Examples

if (FALSE) {
scenario <- readScenario(filename = "scenario.txt")
scenario <- scenario.update.paths(scenario, from = "/home/manuel/", to = "/home/leslie")
}