testConfigurations

testConfigurations executes the given configurations on the testing instances specified in the scenario.

testConfigurations(configurations, scenario, parameters)

Arguments

configurations
(data.frame)
Parameter configurations of the target algorithm (one per row).

scenario
(list())
Data structure containing irace settings. The data structure has to be the one returned by the function defaultScenario or readScenario.

parameters
(list())
Data structure containing the parameter space definition. The data structure has to similar to the one returned by the function readParameters.

Value

A list with the following elements:

experiments

Experiments results.

seeds

Array of the instance seeds used in the experiments.

Details

A test instance set must be provided through scenario$testInstances.

See also