Generate a command-line representation of a configuration

buildCommandLine receives two vectors, one containing the values of the parameters, the other containing the switches of the parameters. It builds a string with the switches and the values that can be used as a command line to call the program to be tuned, thus generating one candidate configuration.

Usage

buildCommandLine(values, switches)

Arguments

values

A vector containing the value of each parameter for the candidate configuration.

switches

A vector containing the switches of each paramter (in an order that corresponds to the values vector).

Value

A string concatenating each element of switches and values for all parameters with a space between each pair of parameters (but none between the switches and the corresponding values).

Author

Manuel López-Ibáñez and Jérémie Dubois-Lacoste

Examples

switches <- c("--switch1 ", "--switch2-", "--switch3=")
values <- list("value_1", 1L, sqrt(2))
buildCommandLine (values, switches)
#> [1] "--switch1 value_1 --switch2-1 --switch3=1.4142135623731"
## Build a command-line from the results produced by a previous run of irace.
# First, load the data produced by irace.
logfile <- file.path(system.file(package="irace"), "exdata", "irace-acotsp.Rdata")
iraceResults <- read_logfile(logfile)
allConfigurations <- iraceResults$allConfigurations
parameters <- iraceResults$scenario$parameters
apply(allConfigurations[1:10, unlist(parameters$names)], 1, buildCommandLine,
      unlist(parameters$switches))
#>                                                                                                                         1 
#>                                      "--as --localsearch 0 --alpha 1.0000 --beta 1.0000 --rho  0.9500 --ants 10 --time 5" 
#>                                                                                                                         2 
#>                    "--as --localsearch 3 --alpha 4.4870 --beta 3.3929 --rho  0.8781 --ants 90 --nnls 38 --dlb 1 --time 5" 
#>                                                                                                                         3 
#> "--eas --localsearch 1 --alpha 1.9870 --beta 8.3929 --rho  0.3831 --ants 20 --nnls 15 --dlb 0 --elitistants 107 --time 5" 
#>                                                                                                                         4 
#>                                    "--mmas --localsearch 0 --alpha 0.7370 --beta 5.8929 --rho  0.6306 --ants 42 --time 5" 
#>                                                                                                                         5 
#>       "--acs --localsearch 2 --alpha 3.2370 --beta 0.8929 --rho  0.1356 --ants  9 --nnls 49 --q0 0.1223 --dlb 0 --time 5" 
#>                                                                                                                         6 
#>                    "--as --localsearch 1 --alpha 2.6120 --beta 4.6429 --rho  0.7543 --ants 29 --nnls  9 --dlb 0 --time 5" 
#>                                                                                                                         7 
#>    "--ras --localsearch 3 --alpha 0.1120 --beta 9.6429 --rho  0.2593 --ants  6 --nnls 32 --dlb 1 --rasranks 100 --time 5" 
#>                                                                                                                         8 
#> "--eas --localsearch 2 --alpha 1.3620 --beta 7.1429 --rho  0.5068 --ants 62 --nnls 43 --dlb 0 --elitistants 201 --time 5" 
#>                                                                                                                         9 
#>                         "--acs --localsearch 0 --alpha 3.8620 --beta 2.1429 --rho  0.0118 --ants 13 --q0 0.7473 --time 5" 
#>                                                                                                                        10 
#>       "--acs --localsearch 3 --alpha 1.6745 --beta 4.0179 --rho  0.6924 --ants 16 --nnls 29 --q0 0.9348 --dlb 1 --time 5"