Generate a command-line representation of a configuration

buildCommandLine receives two vectors, one containing the values of the parameters, the other containing the switches of the parameters. It builds a string with the switches and the values that can be used as a command line to call the program to be tuned, thus generating one candidate configuration.

buildCommandLine(values, switches)

Arguments

values
A vector containing the value of each parameter for the candidate configuration.

switches
A vector containing the switches of each paramter (in an order that corresponds to the values vector).

Value

A string concatenating each element of switches and values for all parameters with a space between each pair of parameters (but none between the switches and the corresponding values).

Examples

switches <- c("--switch1 ", "--switch2 ")
values <- c("value_1", "value_2")
buildCommandLine (values, switches)

#> [1] " --switch1 value_1 --switch2 value_2"

## Build a command-line from the results produced by a previous run of irace.
# First, load the data produced by irace.
irace.logfile <- file.path(system.file(package="irace"),
                           "exdata", "irace-acotsp.Rdata")
load(irace.logfile)
attach(iraceResults)

#> The following object is masked from package:irace:
#> 
#>     irace.version

apply(allConfigurations[1:10, unlist(parameters$names)], 1, buildCommandLine,
      unlist(parameters$switches))

#>                                                                                                           1 
#>                                " --as --localsearch 0 --alpha 1.0000 --beta 1.0000 --rho  0.9500 --ants 10" 
#>                                                                                                           2 
#>                              " --mmas --localsearch 0 --alpha 3.7893 --beta 7.2423 --rho  0.5745 --ants 11" 
#>                                                                                                           3 
#> " --acs --localsearch 2 --alpha 2.3623 --beta 2.1143 --rho  0.0570 --ants  7 --nnls  8 --q0 0.6042 --dlb 0" 
#>                                                                                                           4 
#>                 " --ras --localsearch 0 --alpha 2.7638 --beta 5.5520 --rho  0.1889 --ants 86 --rasranks 41" 
#>                                                                                                           5 
#>                              " --mmas --localsearch 0 --alpha 3.9256 --beta 9.0031 --rho  0.0885 --ants 20" 
#>                                                                                                           6 
#>              " --as --localsearch 3 --alpha 0.1941 --beta 1.5463 --rho  0.5206 --ants  9 --nnls 27 --dlb 1" 
#>                                                                                                           7 
#>            " --mmas --localsearch 3 --alpha 3.2478 --beta 1.8010 --rho  0.9703 --ants 97 --nnls 23 --dlb 1" 
#>                                                                                                           8 
#>            " --mmas --localsearch 2 --alpha 4.7877 --beta 2.9132 --rho  0.6520 --ants 71 --nnls 16 --dlb 0" 
#>                                                                                                           9 
#>                                " --as --localsearch 0 --alpha 3.4702 --beta 6.5742 --rho  0.6527 --ants 54" 
#>                                                                                                          10 
#>            " --mmas --localsearch 3 --alpha 3.4951 --beta 1.8891 --rho  0.9620 --ants 13 --nnls 13 --dlb 1"