Test that the given irace scenario can be run.

checkIraceScenario tests that the given irace scenario can be run by checking the scenario settings provided and trying to run the target-algorithm.

checkIraceScenario(scenario, parameters = NULL)

Arguments

scenario
(list())
Data structure containing irace settings. The data structure has to be the one returned by the function defaultScenario or readScenario.

parameters
(list())
Data structure containing the parameter space definition. The data structure has to similar to the one returned by the function readParameters.

Value

returns TRUE if succesful and gives an error and returns FALSE otherwise.

Details

Provide the parameters argument only if the parameter list should not be obtained from the parameter file given by the scenario. If the parameter list is provided it will not be checked. This function will try to execute the target-algorithm.

See also