Returns the configurations by the iteration in which they were executed.

Returns the configurations by the iteration in which they were executed.

Usage

getConfigurationByIteration(iraceResults, iterations, drop.metadata = FALSE)

Arguments

iraceResults

(list()|character(1))
Object created by irace and typically saved in the log file irace.Rdata. If a character string is given, then it is interpreted as the path to the log file from which the iraceResults object will be loaded.

iterations

(integer())
The iteration number or a vector of iteration numbers from where the configurations should be obtained. Negative values start counting from the last iteration.

drop.metadata

(logical(1))
Remove metadata, such as the configuration ID and the ID of the parent, from the returned configurations. See removeConfigurationsMetaData().

Value

A data frame containing the elite configurations required.

Author

Manuel López-Ibáñez and Leslie Pérez Cáceres

Examples

log_file <- system.file("exdata/irace-acotsp.Rdata", package="irace", mustWork=TRUE)
getConfigurationByIteration(log_file, iterations = c(-2, -1), drop.metadata = TRUE)
#>     algorithm localsearch  alpha   beta    rho ants nnls     q0 dlb rasrank
#> 108       acs           3 1.1286 2.3596 0.4745    6   26 0.2144   1      NA
#> 127       acs           3 0.9926 4.4442 0.4589    7   21 0.4865   1      NA
#> 136       acs           3 1.8399 4.2655 0.2080    5   19 0.0432   1      NA
#> 145       acs           3 1.8844 0.0659 0.2390    7   18 0.3465   1      NA
#> 146       acs           3 1.2254 3.9173 0.0899    7   13 0.2308   1      NA
#> 147       acs           3 1.6397 3.8003 0.4405   10   21 0.6243   1      NA
#> 148       acs           3 1.5729 1.1736 0.4995    5   28 0.0632   1      NA
#> 149       acs           3 1.5633 1.1449 0.2693    8   14 0.3597   1      NA
#> 150       acs           3 0.5250 3.9025 0.4135    5   10 0.5168   1      NA
#> 151       acs           3 1.3612 5.3681 0.3453    5   16 0.5135   1      NA
#> 152       acs           3 1.4320 4.0186 0.2922    6   25 0.0605   1      NA
#> 153       acs           3 2.1014 3.4136 0.3419    5   20 0.0419   1      NA
#> 154       acs           3 1.2471 4.2285 0.4869    6   24 0.1930   1      NA
#> 155       acs           3 2.1420 4.0979 0.1330   13   25 0.3081   1      NA
#> 156       acs           3 1.5956 1.5714 0.4958    5   18 0.0573   1      NA
#> 157       acs           3 0.9501 3.6199 0.5686    5   20 0.3571   1      NA
#> 158       acs           3 1.3778 1.0332 0.6207   10   22 0.1170   1      NA
#> 159       acs           3 0.5481 1.2984 0.4588    5   19 0.2591   1      NA
#> 160       acs           3 0.8909 4.1550 0.4147    8   18 0.4353   1      NA
#> 161       acs           3 0.8896 5.6680 0.3289    6   20 0.6432   1      NA
#> 162       acs           3 0.0158 3.3444 0.3754    5   23 0.4801   1      NA
#> 163       acs           3 1.6732 3.9224 0.3918    5    8 0.3633   1      NA
#>     elitistants
#> 108          NA
#> 127          NA
#> 136          NA
#> 145          NA
#> 146          NA
#> 147          NA
#> 148          NA
#> 149          NA
#> 150          NA
#> 151          NA
#> 152          NA
#> 153          NA
#> 154          NA
#> 155          NA
#> 156          NA
#> 157          NA
#> 158          NA
#> 159          NA
#> 160          NA
#> 161          NA
#> 162          NA
#> 163          NA