Returns the configurations by the iteration in which they were executed.

Returns the configurations by the iteration in which they were executed.

Usage

getConfigurationByIteration(iraceResults, iterations, drop.metadata = FALSE)

Arguments

iraceResults

(list()|character(1))
Object created by irace and typically saved in the log file irace.Rdata. If a character string is given, then it is interpreted as the path to the log file from which the iraceResults object will be loaded.

iterations

(integer())
The iteration number or a vector of iteration numbers from where the configurations should be obtained. Negative values start counting from the last iteration.

drop.metadata

(logical(1))
Remove metadata, such as the configuration ID and the ID of the parent, from the returned configurations. See removeConfigurationsMetaData().

Value

A data frame containing the elite configurations required.

Author

Manuel López-Ibáñez and Leslie Pérez Cáceres

Examples

log_file <- system.file("exdata/irace-acotsp.Rdata", package="irace", mustWork=TRUE)
getConfigurationByIteration(log_file, iterations = c(-2, -1), drop.metadata = TRUE)
#>     algorithm localsearch  alpha   beta    rho ants nnls     q0 dlb rasrank
#> 125       acs           3 2.3822 6.7773 0.1815   22   37 0.0721   1      NA
#> 128       acs           3 1.8725 7.3650 0.4093   17   12 0.2948   1      NA
#> 129       acs           3 2.0116 4.9449 0.6527   24   30 0.1569   1      NA
#> 133       acs           3 2.3234 8.1976 0.8359   51   31 0.6606   1      NA
#> 141       acs           3 1.3378 7.4862 0.3570    7   21 0.1651   1      NA
#> 149       acs           3 1.9696 9.6139 0.1015   44   36 0.0910   1      NA
#> 150       acs           3 2.5703 8.6300 0.6104   18   15 0.2048   1      NA
#> 151       acs           3 2.3764 7.8953 0.6886   47   34 0.5576   1      NA
#> 152       acs           3 0.5587 8.7541 0.3415    8   28 0.1704   1      NA
#> 153       acs           3 1.7734 6.0521 0.1716   21   19 0.4600   1      NA
#> 154       acs           3 4.0355 7.5080 0.7823   66   34 0.7103   1      NA
#> 155       acs           3 1.3706 6.9219 0.1636   22   28 0.0580   1      NA
#> 156       acs           3 0.5064 8.2038 0.1577   11   25 0.0364   1      NA
#> 157       acs           3 2.7098 5.7508 0.1615   17   15 0.2914   1      NA
#> 158       acs           3 1.6219 5.2131 0.2610    6   17 0.0146   1      NA
#> 159       acs           3 1.3815 7.0174 0.4367   19   17 0.1531   1      NA
#>     elitistants
#> 125          NA
#> 128          NA
#> 129          NA
#> 133          NA
#> 141          NA
#> 149          NA
#> 150          NA
#> 151          NA
#> 152          NA
#> 153          NA
#> 154          NA
#> 155          NA
#> 156          NA
#> 157          NA
#> 158          NA
#> 159          NA