Returns the configurations by the iteration in which they were executed.

Returns the configurations by the iteration in which they were executed.

Usage

getConfigurationByIteration(iraceResults, iterations, drop.metadata = FALSE)

Arguments

iraceResults

(list()|character(1))
Object created by irace and typically saved in the log file irace.Rdata. If a character string is given, then it is interpreted as the path to the log file from which the iraceResults object will be loaded.

iterations

(integer())
The iteration number or a vector of iteration numbers from where the configurations should be obtained. Negative values start counting from the last iteration.

drop.metadata

(logical(1))
Remove metadata, such as the configuration ID and the ID of the parent, from the returned configurations. See removeConfigurationsMetaData().

Value

A data frame containing the elite configurations required.

Author

Manuel López-Ibáñez and Leslie Pérez Cáceres

Examples

log_file <- system.file("exdata/irace-acotsp.Rdata", package="irace", mustWork=TRUE)
getConfigurationByIteration(log_file, iterations = c(-2, -1), drop.metadata = TRUE)
#>     algorithm localsearch  alpha   beta    rho ants nnls     q0 dlb rasrank
#> 111       acs           2 1.2881 6.4311 0.5859   10   23 0.2126   0      NA
#> 119       acs           2 1.4432 6.4746 0.4582    5   28 0.1931   0      NA
#> 123       acs           2 1.0584 5.8121 0.5587   10   35 0.2575   0      NA
#> 130       acs           2 1.5899 6.9391 0.5529    8   26 0.3796   0      NA
#> 134       acs           2 1.6144 7.8972 0.1999    8   33 0.1415   1      NA
#> 141       acs           2 1.1411 5.6044 0.7223   16   27 0.3231   0      NA
#> 142       acs           2 1.7492 5.1542 0.6186    7   26 0.1344   0      NA
#> 143       acs           2 0.6967 6.8546 0.6225    8   36 0.0928   0      NA
#> 144       acs           2 0.3108 6.3414 0.6865    7   15 0.0829   0      NA
#> 145       acs           2 1.0345 5.0647 0.4385   10   36 0.2845   0      NA
#> 146       acs           2 1.9401 9.0205 0.6431   21   21 0.2613   0      NA
#> 147       acs           2 0.8712 5.4593 0.3944    7   35 0.2770   1      NA
#> 148       acs           2 0.6575 4.0491 0.3320    8   39 0.2268   1      NA
#>     elitistants
#> 111          NA
#> 119          NA
#> 123          NA
#> 130          NA
#> 134          NA
#> 141          NA
#> 142          NA
#> 143          NA
#> 144          NA
#> 145          NA
#> 146          NA
#> 147          NA
#> 148          NA