Post-selection race

psRace performs a post-selection race of a set of configurations.

Usage

psRace(
  iraceResults,
  max_experiments,
  conf_ids = NULL,
  iteration_elites = FALSE,
  psrace_logFile = NULL
)

Arguments

iraceResults

list()|character(1)
Object created by irace and typically saved in the log file irace.Rdata. If a character string is given, then it is interpreted as the path to the log file from which the iraceResults object will be loaded.

max_experiments

numeric(1)
Number of experiments for the post-selection race. If it is equal to or smaller than 1, then it is a fraction of the total budget given by iraceResults$scenario$maxExperiments or iraceResults$scenario$maxTime / iraceResults$state$boundEstimate.

conf_ids

IDs of the configurations in iraceResults$allConfigurations to be used for the post-selection. If NULL, then the configurations are automatically selected.

iteration_elites

If FALSE, give priority to selecting the configurations that were elite in the last iteration. If TRUE, select from all elite configurations of all iterations. This parameter only has an effect when conf_ids is not NULL.

psrace_logFile

character(1)
Log file to save the post-selection race log. If NULL, the log is saved in iraceResults$scenario$logFile.

Value

The elite configurations after the post-selection. In addition, if iraceResults$scenario$logFile is defined, it saves an updated copy of iraceResults in that file, where iraceResults$psrace_log is a list with the following elements:

configurations

Configurations used in the post-selection race.

instances

Data frame with the instances used in the post-selection race. First column has the instances IDs from iraceResults$scenario$instances, second column the seed assigned to the instance.

max_experiments

Configuration budget assigned to the post-selection race.

experiments

Matrix of results generated by the post-selection race, in the same format as the matrix iraceResults$experiments. Column names are the configuration IDs and row names are the instance IDs.

elites

Best configurations found in the experiments.

Author

Leslie Pérez Cáceres and Manuel López-Ibáñez

Examples

# \donttest{
  irace_log <- read_logfile(system.file(package="irace", "exdata", "sann.rda"))
  # Use a temporary file to not change the original "sann.rda".
  psrace_logFile <- withr::local_tempfile(fileext = ".Rdata")
  # Execute the post-selection after the execution of irace. Use 10% of the total budget.
  psRace(irace_log, max_experiments=0.1, psrace_logFile = psrace_logFile)
#> # 2025-01-04 00:27:45 UTC: Starting post-selection:
#> # scenario maxExperiments:1000
#> # Configurations selected: 120, 127, 119, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13.
#> # Pending instances: 1, 1, 1, 7, 7, 7, 5, 7, 7, 7, 7, 7, 7, 3, 7, 7.
#> # 2025-01-04 00:27:45 UTC: seed: 1355893935
#> # Configurations: 16
#> # Available experiments: 100
#> # minSurvival: 1
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |x|          6|         16|        119|   0.02926503588|         11|00:00:00|   NA|  NA|    NA|
#> |.|          2|         16|        119|   0.04640641217|         11|00:00:00|+0.56|0.78|0.7115|
#> |x|          9|         16|        120|   0.02408726644|         23|00:00:00|+0.43|0.62|0.7491|
#> |.|          5|         16|        120|   0.02198473221|         23|00:00:00|+0.47|0.60|0.7157|
#> |.|          1|         16|        120|    0.6086077084|         23|00:00:00|+0.45|0.56|0.6620|
#> |.|          4|         16|        120|    0.5097405891|         23|00:00:00|+0.45|0.54|0.6327|
#> |x|          7|         16|        127|   0.03721012345|         34|00:00:00|+0.49|0.56|0.6304|
#> |x|          8|         16|        120|    0.3893872037|         46|00:00:00|+0.48|0.54|0.6084|
#> |.|          3|         16|        120|    0.3607054940|         46|00:00:00|+0.49|0.55|0.5991|
#> |x|         10|         16|        119|    0.2654501070|         59|00:00:00|+0.47|0.52|0.5967|
#> |-|         11|          4|        119|    0.2503883469|         72|00:00:00|+0.22|0.29|0.6390|
#> |-|         12|          3|        119|    0.2304387517|         76|00:00:00|-0.08|0.01|0.7244|
#> |=|         13|          3|        119|    0.2134468717|         79|00:00:00|-0.08|0.01|0.7149|
#> |=|         14|          3|        120|    0.8218625008|         82|00:00:00|-0.07|0.01|0.7166|
#> |=|         15|          3|        120|    0.7671186522|         85|00:00:00|-0.06|0.01|0.7111|
#> |=|         16|          3|        120|    0.7192178034|         88|00:00:00|-0.04|0.03|0.7010|
#> |=|         17|          3|        120|    0.6769523820|         91|00:00:00|-0.01|0.04|0.6880|
#> |=|         18|          3|        120|    0.7835383971|         94|00:00:00|-0.02|0.04|0.6888|
#> |=|         19|          3|        120|    0.7522208822|         97|00:00:00|-0.04|0.02|0.6971|
#> |=|         20|          3|        120|    0.7146450673|        100|00:00:00|-0.02|0.03|0.6890|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         120    mean value:     0.7146450673
#> Description of the best-so-far configuration:
#>     .ID. tmax   temp .PARENT.
#> 120  120 2617 1.2212       89
#> 
#> # 2025-01-04 00:27:48 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>     tmax   temp
#> 120 2617 1.2212
#> 127 2553 0.3116
#> 119 2839 1.7558
#> # Total CPU user time: 34.259, CPU sys time: 0.059, Wall-clock time: 34.324
#>     .ID. tmax   temp .PARENT. .RANK.  .WEIGHT.
#> 120  120 2617 1.2212       89     36 0.5000000
#> 127  127 2553 0.3116       89     41 0.3333333
#> 119  119 2839 1.7558       75     43 0.1666667
  # Print psrace_log
  irace_log <- read_logfile(psrace_logFile)
  str(irace_log$psrace_log)
#> List of 5
#>  $ configurations :'data.frame':	3 obs. of  6 variables:
#>   ..$ .ID.    : int [1:3] 120 127 119
#>   ..$ tmax    : int [1:3] 2617 2553 2839
#>   ..$ temp    : num [1:3] 1.221 0.312 1.756
#>   ..$ .PARENT.: int [1:3] 89 89 75
#>   ..$ .RANK.  : num [1:3] 36 41 43
#>   ..$ .WEIGHT.: num [1:3] 0.5 0.333 0.167
#>  $ instances      :Classes ‘data.table’ and 'data.frame':	20 obs. of  2 variables:
#>   ..$ instanceID: int [1:20] 33 46 28 39 2 7 35 1 15 34 ...
#>   ..$ seed      : int [1:20] 1648337051 34104718 614648674 139114586 1017123459 1995230745 294664872 627824735 1136796916 674306437 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ max_experiments: int 100
#>  $ experiments    : num [1:20, 1:16] 1.173 7.796 0.101 3.491 5.529 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:20] "6" "2" "9" "5" ...
#>   .. ..$ : chr [1:16] "1" "2" "3" "4" ...
#>  $ elites         : int [1:3] 120 127 119
# }