Postselection race

psRace performs a postselection race of a set of configurations.

Usage

psRace(
  iraceResults,
  max_experiments,
  conf_ids = NULL,
  iteration_elites = FALSE
)

Arguments

iraceResults

list()|character(1)
Object created by irace and typically saved in the log file irace.Rdata. If a character string is given, then it is interpreted as the path to the log file from which the iraceResults object will be loaded.

max_experiments

numeric(1)
Number of experiments for the post-selection race. If it is equal to or smaller than 1, then it is a fraction of the total budget given by iraceResults$scenario$maxExperiments or iraceResults$scenario$maxTime / iraceResults$state$boundEstimate.

conf_ids

IDs of the configurations in iraceResults$allConfigurations to be used for ablation. If NULL, the iteration_best argument will be to decide.

iteration_elites

If TRUE, only select the best configuration of each iteration. If FALSE, select from all elite configurations of all iterations. max_experiments

Value

If iraceLogFile is NULL, it returns a list with the following elements:

configurations

Configurations used in the race.

instances

A matrix with the instances used in the experiments. First column has the instances ids from iraceResults$scenario$instances, second column the seed assigned to the instance.

maxExperiments

Maximum number of experiments set for the race.

experiments

A matrix with the results of the experiments (columns are configurations, rows are instances).

elites

Best configurations found in the experiments.

If iraceLogFile is provided this list object will be saved in iraceResults$psrace_log.

Author

Leslie Pérez Cáceres and Manuel López-Ibáñez

Examples

# \donttest{
  logfile <- system.file(package="irace", "exdata", "sann.rda")
  # Execute the postselection after the execution of irace. Use 10% of the total budget.
  psRace(logfile, max_experiments=0.1)
#> # 2024-12-09 18:28:13 UTC: Starting post-selection:
#> # scenario maxExperiments:1000
#> # Configurations selected: 114, 94, 77, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143, 144, 145, 146.
#> # Pending instances: 0, 0, 0, 3, 3, 3, 4, 4, 3, 3, 3, 3, 3, 3, 3, 3, 4, 3, 4, 3, 3, 3, 3, 3, 4, 4, 4, 4, 3, 2, 4, 3, 4, 4, 4, 3, 3, 3, 3, 4, 2, 4, 3, 3, 4, 4, 4, 3, 4, 2, 3, 4, 3, 4, 4, 3, 3, 4, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 0, 2, 2, 2, 2, 2, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1.
#> # Seed: 1945603016
#> # Configurations: 146
#> # Available experiments: 100
#> # minSurvival: 1
#> # Markers:
#>      x No test is performed.
#>      c Configurations are discarded only due to capping.
#>      - The test is performed and some configurations are discarded.
#>      = The test is performed but no configuration is discarded.
#>      ! The test is performed and configurations could be discarded but elite configurations are preserved.
#>      . All alive configurations are elite and nothing is discarded.
#> 
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> | |   Instance|      Alive|       Best|       Mean best| Exp so far|  W time|  rho|KenW|  Qvar|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> |.|          4|        146|         69|  0.008691232201|          0|00:00:00|   NA|  NA|    NA|
#> |.|          3|        146|         69|     1.541677024|          0|00:00:00|+0.19|0.59|0.9515|
#> |x|          8|        146|        110|   0.04819047679|        101|00:00:02|+0.20|0.46|0.8554|
#> |x|          9|        146|        114|   0.04809467679|        240|00:00:03|+0.28|0.46|0.8051|
#> |!|          7|        146|        114|   0.04825551260|        292|00:00:01|+0.28|0.42|0.7955|
#> |.|          2|        146|        114|   0.05968208546|        292|00:00:00|+0.27|0.39|0.7938|
#> |.|          5|        146|        114|   0.06010600025|        292|00:00:00|+0.23|0.34|0.8279|
#> +-+-----------+-----------+-----------+----------------+-----------+--------+-----+----+------+
#> Best-so-far configuration:         114    mean value:    0.06532854247
#> Description of the best-so-far configuration:
#>     .ID. tmax    temp .PARENT.
#> 114  114  330 35.7786       94
#> 
#> # 2024-12-09 18:28:20 UTC: Elite configurations (first number is the configuration ID; listed from best to worst according to the sum of ranks):
#>     tmax    temp
#> 114  330 35.7786
#> 145  196 35.1518
#> 77   368 45.3900
#> # Total CPU user time: 35.66, CPU sys time: 0.012, Wall-clock time: 35.691
#>     .ID. tmax    temp .PARENT. .RANK.  .WEIGHT.
#> 114  114  330 35.7786       94    165 0.5000000
#> 145  145  196 35.1518       94    273 0.3333333
#> 77    77  368 45.3900       27    284 0.1666667
# }