psRace

psRace performs a postselection race a set of configurations.

psRace(iraceLogFile = NULL, iraceResults = NULL, conf.ids = NULL,
  postselection = NULL, max.experiments = NULL, elites = FALSE,
  seed = 1234567)

Arguments

iraceLogFile
NULL Log file created by irace, this file must contain the iraceResults object.

iraceResults
NULL Object created by irace and saved in scenario$logFile.

conf.ids
NULL IDs of the configurations in iraceResults$allConfigurations to be used for ablation. If NULL, the elites argument will be used.

postselection
NULL Percentage of the maxExperiments provided in the scenario to be used in the race.

max.experiments
NULL Number of experiments available for the race. If NULL budget for the race is set by the parameter scenario$postselection, which defines the percentage of the total budget of irace (iraceResults$scenario$maxExperiments or iraceResults$scenario$maxTime/iraceResults$state$timeEstimate) to use for the postselection.

elites
FALSE Flag for selecting configurations. If FALSE, the best configurations of each iteration are used for the race. If TRUE, the elite configurtions of each iteration are used for the race.

seed
1234567 Numerical value to use as seed for the random number generation.

Value

If iraceLogFile is NULL, it returns a list with the following elements:

configurations

Configurations used in the race.

instances

A matrix with the instances used in the experiments. First column has the instances ids from iraceResults$scenario$instances, second column the seed assigned to the instance.

maxExperiments

Maximum number of experiments set for the race.

experiments

A matrix with the results of the experiments (columns are configurations, rows are instances).

elites

Best configurations found in the experiments.

If iraceLogFile is provided this list object will be saved in iraceResults$psrace.log.

Examples

# \dontrun{
  # Execute the postselection automatically after irace
  parameters <- readParameters("parameters.txt")

#> Error: == irace == readParameter: parameter file '/tmp/Rtmpz6PKMt/file644f58007edb/reference/parameters.txt' does not exist

  scenario <- readScenario(filename="scenario.txt",
                           scenario=defaultScenario())

#> Error: == irace == The scenario file 'scenario.txt' does not exist.

  # Use 10% of the total budget
  scenario$postselection <- 0.1
  irace(scenario=scenario, parameters=parameters)

#> Error: == irace == target runner 'targetRunner' (--target-runner) '/tmp/Rtmpz6PKMt/file644f58007edb/reference/target-runner' does not exist

  # Execute the postselection after the execution of \pkg{irace}.
  psRace(iraceLogFile="irace.Rdata", max.experiments=120)

#> # 2020-03-30 21:22:18 UTC: Stating post-selection:
#> # Seed:1234567
#> # Log file: irace.Rdata 

#> Warning: cannot open compressed file 'irace.Rdata', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

# }