Read parameter configurations from a file

Reads a set of target-algorithm configurations from a file and puts them in irace format. The configurations are checked to match the parameters description provided.

Usage

readConfigurationsFile(filename, parameters, debugLevel = 0, text)

Arguments

filename

(character(1))
Filename from which the configurations should be read. The contents should be readable by read.table( , header=TRUE).

parameters

(list())
Data structure containing the parameter space definition. The data structure has to similar to the one returned by the function readParameters.

debugLevel

(integer(1))
Larger values produce more verbose output.

text

(character(1))
If file is not supplied and this is, then parameters are read from the value of text via a text connection.

Value

A data frame containing the obtained configurations. Each row of the data frame is a candidate configuration, the columns correspond to the parameter names in parameters.

Details

Example of an input file:

# This is a comment line
<param_name_1>  <param_name_2> ... 
      0.5           "value_1"    ...
      1.0           "value_2"    ...
      1.2           "value_3"    ...
      ...           ...

The order of the columns does not necessarily have to be the same as in the file containing the definition of the parameters.

See also

Author

Manuel López-Ibáñez and Jérémie Dubois-Lacoste