Read the log file produced by irace (irace.Rdata).

Read the log file produced by irace (irace.Rdata).

Usage

read_logfile(filename, name = "iraceResults")

Arguments

filename

Filename that contains the log file saved by irace. Example: irace.Rdata.

name

Optional argument that allows overriding the default name of the object in the file.

Value

(list())

Examples

irace_results <- read_logfile(system.file("exdata/irace-acotsp.Rdata", package="irace",
                                          mustWork=TRUE))
str(irace_results)
#> List of 11
#>  $ scenario         :List of 58
#>   ..$ scenarioFile             : chr "/home/manu/work/irace/git/devel-examples/vignette-example/scenario.txt"
#>   ..$ configurationsFile       : chr "/home/manu/work/irace/git/devel-examples/vignette-example/default.txt"
#>   ..$ logFile                  : chr "./irace-acotsp.Rdata"
#>   ..$ trainInstancesDir        : chr "./instances"
#>   ..$ trainInstancesFile       : chr "./instances.txt"
#>   ..$ testInstancesDir         : chr "./instances"
#>   ..$ testInstancesFile        : chr "./instances-test.txt"
#>   ..$ testNbElites             : int 5
#>   ..$ testIterationElites      : logi TRUE
#>   ..$ maxExperiments           : int 1000
#>   ..$ seed                     : int 687542627
#>   ..$ execDir                  : chr "./"
#>   ..$ parameterFile            : chr "./parameters.txt"
#>   ..$ recoveryFile             : chr ""
#>   ..$ instances                : chr [1:10] "./instances/1000-1.tsp" "./instances/1000-2.tsp" "./instances/1000-3.tsp" "./instances/1000-4.tsp" ...
#>   ..$ sampleInstances          : logi TRUE
#>   ..$ testInstances            : chr [1:10] "./instances/1000-10.tsp" "./instances/1000-11.tsp" "./instances/1000-12.tsp" "./instances/1000-13.tsp" ...
#>   ..$ testType                 : chr "friedman"
#>   ..$ firstTest                : int 5
#>   ..$ blockSize                : int 1
#>   ..$ eachTest                 : int 1
#>   ..$ targetRunner             : chr "./target-runner"
#>   ..$ targetRunnerLauncher     : chr ""
#>   ..$ targetCmdline            : chr "{configurationID} {instanceID} {seed} {instance} {bound} {targetRunnerArgs}"
#>   ..$ targetRunnerRetries      : int 0
#>   ..$ targetRunnerTimeout      : int 0
#>   ..$ targetRunnerData         : chr ""
#>   ..$ deterministic            : logi FALSE
#>   ..$ minExperiments           : chr NA
#>   ..$ maxTime                  : int 0
#>   ..$ budgetEstimation         : num 0.02
#>   ..$ minMeasurableTime        : num 0.01
#>   ..$ parallel                 : int 2
#>   ..$ loadBalancing            : logi TRUE
#>   ..$ mpi                      : logi FALSE
#>   ..$ batchmode                : chr "0"
#>   ..$ digits                   : int 4
#>   ..$ quiet                    : logi FALSE
#>   ..$ debugLevel               : int 0
#>   ..$ softRestart              : logi TRUE
#>   ..$ softRestartThreshold     : num 1e-04
#>   ..$ elitist                  : logi TRUE
#>   ..$ elitistNewInstances      : int 1
#>   ..$ elitistLimit             : int 2
#>   ..$ capping                  : logi FALSE
#>   ..$ cappingType              : chr "median"
#>   ..$ boundType                : chr "candidate"
#>   ..$ boundDigits              : int 0
#>   ..$ boundPar                 : int 1
#>   ..$ boundAsTimeout           : logi TRUE
#>   ..$ postselection            : num 0
#>   ..$ aclib                    : logi FALSE
#>   ..$ nbIterations             : int 0
#>   ..$ nbExperimentsPerIteration: int 0
#>   ..$ minNbSurvival            : int 0
#>   ..$ nbConfigurations         : int 0
#>   ..$ mu                       : int 5
#>   ..$ confidence               : num 0.95
#>  $ irace.version    : chr "3.5.1.9000.0d7b9f8-dirty"
#>  $ parameters       :List of 15
#>   ..$ names       : chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ types       : Named chr [1:11] "c" "c" "r" "r" ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ switches    : Named chr [1:11] "--" "--localsearch " "--alpha " "--beta " ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ domain      :List of 11
#>   .. ..$ algorithm  : chr [1:5] "as" "mmas" "eas" "ras" ...
#>   .. ..$ localsearch: chr [1:4] "0" "1" "2" "3"
#>   .. ..$ alpha      : num [1:2] 0 5
#>   .. ..$ beta       : num [1:2] 0 10
#>   .. ..$ rho        : num [1:2] 0.01 1
#>   .. ..$ ants       : num [1:2] 5 100
#>   .. ..$ nnls       : num [1:2] 5 50
#>   .. ..$ q0         : num [1:2] 0 1
#>   .. ..$ dlb        : chr [1:2] "0" "1"
#>   .. ..$ rasrank    : num [1:2] 1 100
#>   .. ..$ elitistants: num [1:2] 1 750
#>   ..$ conditions  :List of 11
#>   .. ..$ algorithm  : logi TRUE
#>   .. ..$ localsearch: logi TRUE
#>   .. ..$ alpha      : logi TRUE
#>   .. ..$ beta       : logi TRUE
#>   .. ..$ rho        : logi TRUE
#>   .. ..$ ants       : logi TRUE
#>   .. ..$ nnls       :  expression(localsearch %in% c(1, 2, 3))
#>   .. ..$ q0         :  expression(algorithm %in% c("acs"))
#>   .. ..$ dlb        :  expression(localsearch %in% c(1, 2, 3))
#>   .. ..$ rasrank    :  expression(algorithm %in% c("ras"))
#>   .. ..$ elitistants:  expression(algorithm %in% c("eas"))
#>   ..$ isFixed     : Named logi [1:11] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ transform   :List of 11
#>   .. ..$ algorithm  : chr ""
#>   .. ..$ localsearch: chr ""
#>   .. ..$ alpha      : chr ""
#>   .. ..$ beta       : chr ""
#>   .. ..$ rho        : chr ""
#>   .. ..$ ants       : chr "log"
#>   .. .. ..- attr(*, "lower")= num 1.61
#>   .. .. ..- attr(*, "upper")= num 4.62
#>   .. ..$ nnls       : chr ""
#>   .. ..$ q0         : chr ""
#>   .. ..$ dlb        : chr ""
#>   .. ..$ rasrank    : chr ""
#>   .. ..$ elitistants: chr ""
#>   ..$ isDependent : Named logi [1:11] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ digits      : Named int [1:4] 4 4 4 4
#>   .. ..- attr(*, "names")= chr [1:4] "alpha" "beta" "rho" "q0"
#>   ..$ forbidden   :List of 1
#>   .. ..$ :<bytecode> 
#>   .. .. ..- attr(*, "source")= chr "(alpha == 0) & (beta == 0)"
#>   ..$ depends     :List of 11
#>   .. ..$ algorithm  : chr(0) 
#>   .. ..$ localsearch: chr(0) 
#>   .. ..$ alpha      : chr(0) 
#>   .. ..$ beta       : chr(0) 
#>   .. ..$ rho        : chr(0) 
#>   .. ..$ ants       : chr(0) 
#>   .. ..$ nnls       : chr "localsearch"
#>   .. ..$ q0         : chr "algorithm"
#>   .. ..$ dlb        : chr "localsearch"
#>   .. ..$ rasrank    : chr "algorithm"
#>   .. ..$ elitistants: chr "algorithm"
#>   ..$ hierarchy   : Named num [1:11] 1 1 1 1 1 1 2 2 2 2 ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ nbParameters: int 11
#>   ..$ nbFixed     : int 0
#>   ..$ nbVariable  : int 11
#>  $ allElites        :List of 7
#>   ..$ : int [1:4] 2 29 3 20
#>   ..$ : int [1:5] 50 29 2 34 47
#>   ..$ : int [1:5] 50 29 86 47 74
#>   ..$ : int [1:5] 111 50 106 74 92
#>   ..$ : int [1:5] 123 134 130 111 119
#>   ..$ : int [1:5] 119 134 111 130 123
#>   ..$ : int [1:5] 123 119 111 130 134
#>  $ experiments      : num [1:17, 1:148] 28407696 28402347 29166659 28417102 28823388 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:17] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:148] "1" "2" "3" "4" ...
#>  $ experimentLog    : int [1:988, 1:5] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "iteration" "instance" "configuration" "time" ...
#>  $ state            :List of 18
#>   ..$ .Random.seed        : int [1:626] 10403 524 809639586 1907665986 742392805 195146023 -374520492 163985276 951571335 -1799089973 ...
#>   ..$ .irace              :<environment: 0x555ed80902f8> 
#>   ..$ currentBudget       : num 44
#>   ..$ debugLevel          : int 0
#>   ..$ eliteConfigurations :'data.frame':	5 obs. of  16 variables:
#>   .. ..$ .ID.       : int [1:5] 123 119 111 130 134
#>   .. ..$ algorithm  : chr [1:5] "acs" "acs" "acs" "acs" ...
#>   .. ..$ localsearch: chr [1:5] "2" "2" "2" "2" ...
#>   .. ..$ alpha      : num [1:5] 1.06 1.44 1.29 1.59 1.61
#>   .. ..$ beta       : num [1:5] 5.81 6.47 6.43 6.94 7.9
#>   .. ..$ rho        : num [1:5] 0.559 0.458 0.586 0.553 0.2
#>   .. ..$ ants       : num [1:5] 10 5 10 8 8
#>   .. ..$ nnls       : num [1:5] 35 28 23 26 33
#>   .. ..$ q0         : num [1:5] 0.258 0.193 0.213 0.38 0.141
#>   .. ..$ dlb        : chr [1:5] "0" "0" "0" "0" ...
#>   .. ..$ rasrank    : num [1:5] NA NA NA NA NA
#>   .. ..$ elitistants: num [1:5] NA NA NA NA NA
#>   .. ..$ .PARENT.   : int [1:5] 111 111 47 111 92
#>   .. ..$ .ALIVE.    : logi [1:5] TRUE TRUE TRUE TRUE TRUE
#>   .. ..$ .RANK.     : num [1:5] 55 56 58 61 61
#>   .. ..$ .WEIGHT.   : num [1:5] 0.3333 0.2667 0.2 0.1333 0.0667
#>   ..$ experimentsUsedSoFar: int 988
#>   ..$ indexIteration      : int 8
#>   ..$ minSurvival         : num 5
#>   ..$ model               :List of 11
#>   .. ..$ algorithm  :List of 5
#>   .. .. ..$ 119: num [1:5] 0.000268 0.000268 0.000268 0.000604 0.998592
#>   .. .. ..$ 134: num [1:5] 0.000268 0.000268 0.000268 0.000604 0.998592
#>   .. .. ..$ 111: num [1:5] 0.000268 0.000268 0.000268 0.000604 0.998592
#>   .. .. ..$ 130: num [1:5] 0.000268 0.000268 0.000268 0.000604 0.998592
#>   .. .. ..$ 123: num [1:5] 0.000268 0.000268 0.000268 0.000604 0.998592
#>   .. ..$ localsearch:List of 5
#>   .. .. ..$ 119: num [1:4] 0.000336 0.000336 0.998655 0.000672
#>   .. .. ..$ 134: num [1:4] 0.000336 0.000336 0.998655 0.000672
#>   .. .. ..$ 111: num [1:4] 0.000336 0.000336 0.998655 0.000672
#>   .. .. ..$ 130: num [1:4] 0.000336 0.000336 0.998655 0.000672
#>   .. .. ..$ 123: num [1:4] 0.000336 0.000336 0.998655 0.000672
#>   .. ..$ alpha      :List of 5
#>   .. .. ..$ 119: num [1:2] 0.604 1.443
#>   .. .. ..$ 134: num [1:2] 0.604 1.614
#>   .. .. ..$ 111: num [1:2] 0.604 1.288
#>   .. .. ..$ 130: num [1:2] 0.604 1.59
#>   .. .. ..$ 123: num [1:2] 0.604 1.058
#>   .. ..$ beta       :List of 5
#>   .. .. ..$ 119: num [1:2] 1.21 6.47
#>   .. .. ..$ 134: num [1:2] 1.21 7.9
#>   .. .. ..$ 111: num [1:2] 1.21 6.43
#>   .. .. ..$ 130: num [1:2] 1.21 6.94
#>   .. .. ..$ 123: num [1:2] 1.21 5.81
#>   .. ..$ rho        :List of 5
#>   .. .. ..$ 119: num [1:2] 0.12 0.458
#>   .. .. ..$ 134: num [1:2] 0.12 0.2
#>   .. .. ..$ 111: num [1:2] 0.12 0.586
#>   .. .. ..$ 130: num [1:2] 0.12 0.553
#>   .. .. ..$ 123: num [1:2] 0.12 0.559
#>   .. ..$ ants       :List of 5
#>   .. .. ..$ 119: num [1:2] 0.121 5
#>   .. .. ..$ 134: num [1:2] 0.121 8
#>   .. .. ..$ 111: num [1:2] 0.121 10
#>   .. .. ..$ 130: num [1:2] 0.121 8
#>   .. .. ..$ 123: num [1:2] 0.121 10
#>   .. ..$ nnls       :List of 5
#>   .. .. ..$ 119: num [1:2] 5.43 28
#>   .. .. ..$ 134: num [1:2] 5.43 33
#>   .. .. ..$ 111: num [1:2] 5.43 23
#>   .. .. ..$ 130: num [1:2] 5.43 26
#>   .. .. ..$ 123: num [1:2] 5.43 35
#>   .. ..$ q0         :List of 5
#>   .. .. ..$ 119: num [1:2] 0.121 0.193
#>   .. .. ..$ 134: num [1:2] 0.121 0.141
#>   .. .. ..$ 111: num [1:2] 0.121 0.213
#>   .. .. ..$ 130: num [1:2] 0.121 0.38
#>   .. .. ..$ 123: num [1:2] 0.121 0.258
#>   .. ..$ dlb        :List of 5
#>   .. .. ..$ 119: num [1:2] 0.999282 0.000718
#>   .. .. ..$ 134: num [1:2] 0.0262 0.9738
#>   .. .. ..$ 111: num [1:2] 0.999282 0.000718
#>   .. .. ..$ 130: num [1:2] 0.999282 0.000718
#>   .. .. ..$ 123: num [1:2] 0.999282 0.000718
#>   .. ..$ rasrank    :List of 5
#>   .. .. ..$ 119: num [1:2] 12 13
#>   .. .. ..$ 134: num [1:2] 12 13
#>   .. .. ..$ 111: num [1:2] 12 13
#>   .. .. ..$ 130: num [1:2] 12 13
#>   .. .. ..$ 123: num [1:2] 12 13
#>   .. ..$ elitistants:List of 5
#>   .. .. ..$ 119: num [1:2] 90.5 NA
#>   .. .. ..$ 134: num [1:2] 90.5 NA
#>   .. .. ..$ 111: num [1:2] 90.5 NA
#>   .. .. ..$ 130: num [1:2] 90.5 NA
#>   .. .. ..$ 123: num [1:2] 90.5 NA
#>   ..$ nbConfigurations    : num 7
#>   ..$ nbIterations        : int 7
#>   ..$ remainingBudget     : int 12
#>   ..$ timeUsed            : num 0
#>   ..$ boundEstimate       : logi NA
#>   ..$ rejectedIDs         : NULL
#>   ..$ completed           : chr "Not enough budget to race more than the minimum configurations"
#>   ..$ elapsed             : Named num [1:3] 9941.8 63.2 5245.7
#>   .. ..- attr(*, "names")= chr [1:3] "user" "system" "wallclock"
#>   ..$ sessionInfo         :List of 13
#>   .. ..$ R.version      :List of 14
#>   .. .. ..$ platform      : chr "x86_64-pc-linux-gnu"
#>   .. .. ..$ arch          : chr "x86_64"
#>   .. .. ..$ os            : chr "linux-gnu"
#>   .. .. ..$ system        : chr "x86_64, linux-gnu"
#>   .. .. ..$ status        : chr ""
#>   .. .. ..$ major         : chr "4"
#>   .. .. ..$ minor         : chr "1.2"
#>   .. .. ..$ year          : chr "2021"
#>   .. .. ..$ month         : chr "11"
#>   .. .. ..$ day           : chr "01"
#>   .. .. ..$ svn rev       : chr "81115"
#>   .. .. ..$ language      : chr "R"
#>   .. .. ..$ version.string: chr "R version 4.1.2 (2021-11-01)"
#>   .. .. ..$ nickname      : chr "Bird Hippie"
#>   .. ..$ platform       : chr "x86_64-pc-linux-gnu (64-bit)"
#>   .. ..$ locale         : chr "LC_CTYPE=en_GB.UTF-8;LC_NUMERIC=C;LC_TIME=en_GB.UTF-8;LC_COLLATE=en_GB.UTF-8;LC_MONETARY=en_GB.UTF-8;LC_MESSAGE"| __truncated__
#>   .. ..$ running        : chr "Ubuntu 22.04.3 LTS"
#>   .. ..$ RNGkind        : chr [1:3] "Mersenne-Twister" "Inversion" "Rejection"
#>   .. ..$ basePkgs       : chr [1:7] "stats" "graphics" "grDevices" "utils" ...
#>   .. ..$ otherPkgs      :List of 1
#>   .. .. ..$ irace:List of 24
#>   .. .. .. ..$ Package                : chr "irace"
#>   .. .. .. ..$ Type                   : chr "Package"
#>   .. .. .. ..$ Title                  : chr "Iterated Racing for Automatic Algorithm Configuration"
#>   .. .. .. ..$ Description            : chr "Iterated race is an extension of the Iterated F-race method for\n             the automatic configuration of op"| __truncated__
#>   .. .. .. ..$ Version                : chr "3.5.1.9000"
#>   .. .. .. ..$ Authors@R              : chr "c(person(\"Manuel\", \"López-Ibáñez\", role = c(\"aut\", \"cre\"),\n                              email = \"man"| __truncated__
#>   .. .. .. ..$ Depends                : chr "R (>= 3.6.0)"
#>   .. .. .. ..$ Imports                : chr "stats, utils, compiler, matrixStats, R6, withr"
#>   .. .. .. ..$ Suggests               : chr "Rmpi (>= 0.6.0), parallel, knitr, testthat (>= 3.1.8),\nParamHelpers, devtools"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ License                : chr "GPL (>= 2)"
#>   .. .. .. ..$ URL                    : chr "https://mlopez-ibanez.github.io/irace/,\nhttps://github.com/MLopez-Ibanez/irace"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/MLopez-Ibanez/irace/issues"
#>   .. .. .. ..$ ByteCompile            : chr "yes"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ Roxygen                : chr "list(markdown = TRUE)"
#>   .. .. .. ..$ RoxygenNote            : chr "7.2.3"
#>   .. .. .. ..$ SystemRequirements     : chr "GNU make"
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ NeedsCompilation       : chr "yes"
#>   .. .. .. ..$ Packaged               : chr "2023-10-01 11:07:43 UTC; manu"
#>   .. .. .. ..$ Author                 : chr "Manuel López-Ibáñez [aut, cre]\n    (<https://orcid.org/0000-0001-9974-1295>),\n  Jérémie Dubois-Lacoste [aut],"| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Manuel López-Ibáñez <manuel.lopez-ibanez@manchester.ac.uk>"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; x86_64-pc-linux-gnu; 2023-10-01 11:07:44 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/irace/Meta/package.rds"
#>   .. ..$ loadedOnly     :List of 5
#>   .. .. ..$ compiler   :List of 10
#>   .. .. .. ..$ Package    : chr "compiler"
#>   .. .. .. ..$ Version    : chr "4.1.2"
#>   .. .. .. ..$ Priority   : chr "base"
#>   .. .. .. ..$ Title      : chr "The R Compiler Package"
#>   .. .. .. ..$ Author     : chr "Luke Tierney <luke-tierney@uiowa.edu>"
#>   .. .. .. ..$ Maintainer : chr "R Core Team <do-use-Contact-address@r-project.org>"
#>   .. .. .. ..$ Contact    : chr "R-help mailing list <r-help@r-project.org>"
#>   .. .. .. ..$ Description: chr "Byte code compiler for R."
#>   .. .. .. ..$ License    : chr "Part of R 4.1.2"
#>   .. .. .. ..$ Built      : chr "R 4.1.2; ; 2022-02-09 05:09:18 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/library/compiler/Meta/package.rds"
#>   .. .. ..$ R6         :List of 18
#>   .. .. .. ..$ Package         : chr "R6"
#>   .. .. .. ..$ Title           : chr "Encapsulated Classes with Reference Semantics"
#>   .. .. .. ..$ Version         : chr "2.5.1"
#>   .. .. .. ..$ Authors@R       : chr "person(\"Winston\", \"Chang\", role = c(\"aut\", \"cre\"), email = \"winston@stdout.org\")"
#>   .. .. .. ..$ Description     : chr "Creates classes with reference semantics, similar to R's built-in\n    reference classes. Compared to reference"| __truncated__
#>   .. .. .. ..$ Depends         : chr "R (>= 3.0)"
#>   .. .. .. ..$ Suggests        : chr "testthat, pryr"
#>   .. .. .. ..$ License         : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL             : chr "https://r6.r-lib.org, https://github.com/r-lib/R6/"
#>   .. .. .. ..$ BugReports      : chr "https://github.com/r-lib/R6/issues"
#>   .. .. .. ..$ RoxygenNote     : chr "7.1.1"
#>   .. .. .. ..$ NeedsCompilation: chr "no"
#>   .. .. .. ..$ Packaged        : chr "2021-08-06 20:18:46 UTC; winston"
#>   .. .. .. ..$ Author          : chr "Winston Chang [aut, cre]"
#>   .. .. .. ..$ Maintainer      : chr "Winston Chang <winston@stdout.org>"
#>   .. .. .. ..$ Repository      : chr "CRAN"
#>   .. .. .. ..$ Date/Publication: chr "2021-08-19 14:00:05 UTC"
#>   .. .. .. ..$ Built           : chr "R 4.1.1; ; 'Sat, 21 Aug 2021 18:04:39 +0000'; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/site-library/R6/Meta/package.rds"
#>   .. .. ..$ parallel   :List of 14
#>   .. .. .. ..$ Package         : chr "parallel"
#>   .. .. .. ..$ Version         : chr "4.1.2"
#>   .. .. .. ..$ Priority        : chr "base"
#>   .. .. .. ..$ Title           : chr "Support for Parallel computation in R"
#>   .. .. .. ..$ Author          : chr "R Core Team"
#>   .. .. .. ..$ Maintainer      : chr "R Core Team <do-use-Contact-address@r-project.org>"
#>   .. .. .. ..$ Contact         : chr "R-help mailing list <r-help@r-project.org>"
#>   .. .. .. ..$ Description     : chr "Support for parallel computation, including by forking\n   (taken from package multicore), by sockets (taken fr"| __truncated__
#>   .. .. .. ..$ License         : chr "Part of R 4.1.2"
#>   .. .. .. ..$ Imports         : chr "tools, compiler"
#>   .. .. .. ..$ Suggests        : chr "methods"
#>   .. .. .. ..$ Enhances        : chr "snow, nws, Rmpi"
#>   .. .. .. ..$ NeedsCompilation: chr "yes"
#>   .. .. .. ..$ Built           : chr "R 4.1.2; x86_64-pc-linux-gnu; 2022-02-09 05:09:20 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/library/parallel/Meta/package.rds"
#>   .. .. ..$ withr      :List of 24
#>   .. .. .. ..$ Package                : chr "withr"
#>   .. .. .. ..$ Title                  : chr "Run Code 'With' Temporarily Modified Global State"
#>   .. .. .. ..$ Version                : chr "2.5.0"
#>   .. .. .. ..$ Authors@R              : chr "\n    c(person(given = \"Jim\",\n             family = \"Hester\",\n             role = \"aut\"),\n      person"| __truncated__
#>   .. .. .. ..$ Description            : chr "A set of functions to run code 'with' safely and\n    temporarily modified global state. Many of these function"| __truncated__
#>   .. .. .. ..$ License                : chr "MIT + file LICENSE"
#>   .. .. .. ..$ URL                    : chr "https://withr.r-lib.org, https://github.com/r-lib/withr#readme"
#>   .. .. .. ..$ BugReports             : chr "https://github.com/r-lib/withr/issues"
#>   .. .. .. ..$ Depends                : chr "R (>= 3.2.0)"
#>   .. .. .. ..$ Imports                : chr "graphics, grDevices, stats"
#>   .. .. .. ..$ Suggests               : chr "callr, covr, DBI, knitr, lattice, methods, rlang, rmarkdown\n(>= 2.12), RSQLite, testthat (>= 3.0.0)"
#>   .. .. .. ..$ VignetteBuilder        : chr "knitr"
#>   .. .. .. ..$ Encoding               : chr "UTF-8"
#>   .. .. .. ..$ RoxygenNote            : chr "7.1.2"
#>   .. .. .. ..$ Collate                : chr "'aaa.R' 'collate.R' 'compat-defer.R' 'connection.R' 'db.R'\n'defer.R' 'wrap.R' 'local_.R' 'with_.R' 'devices.R'"| __truncated__
#>   .. .. .. ..$ Config/testthat/edition: chr "3"
#>   .. .. .. ..$ Config/Needs/website   : chr "tidyverse/tidytemplate"
#>   .. .. .. ..$ NeedsCompilation       : chr "no"
#>   .. .. .. ..$ Packaged               : chr "2022-03-03 21:13:01 UTC; lionel"
#>   .. .. .. ..$ Author                 : chr "Jim Hester [aut],\n  Lionel Henry [aut, cre],\n  Kirill Müller [aut],\n  Kevin Ushey [aut],\n  Hadley Wickham ["| __truncated__
#>   .. .. .. ..$ Maintainer             : chr "Lionel Henry <lionel@rstudio.com>"
#>   .. .. .. ..$ Repository             : chr "CRAN"
#>   .. .. .. ..$ Date/Publication       : chr "2022-03-03 21:50:02 UTC"
#>   .. .. .. ..$ Built                  : chr "R 4.1.2; ; 2023-03-08 18:52:18 UTC; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/home/manu/R/x86_64-pc-linux-gnu-library/4.1/withr/Meta/package.rds"
#>   .. .. ..$ matrixStats:List of 21
#>   .. .. .. ..$ Package         : chr "matrixStats"
#>   .. .. .. ..$ Version         : chr "0.61.0"
#>   .. .. .. ..$ Depends         : chr "R (>= 2.12.0)"
#>   .. .. .. ..$ Suggests        : chr "base64enc, ggplot2, knitr, markdown, microbenchmark,\nR.devices, R.rsp"
#>   .. .. .. ..$ VignetteBuilder : chr "R.rsp"
#>   .. .. .. ..$ Title           : chr "Functions that Apply to Rows and Columns of Matrices (and to\nVectors)"
#>   .. .. .. ..$ Authors@R       : chr "c(\n  person(\"Henrik\", \"Bengtsson\", role=c(\"aut\", \"cre\", \"cph\"),\n                                 em"| __truncated__
#>   .. .. .. ..$ Author          : chr "Henrik Bengtsson [aut, cre, cph],\n  Constantin Ahlmann-Eltze [ctb],\n  Hector Corrada Bravo [ctb],\n  Robert G"| __truncated__
#>   .. .. .. ..$ Maintainer      : chr "Henrik Bengtsson <henrikb@braju.com>"
#>   .. .. .. ..$ Description     : chr "High-performing functions operating on rows and columns of matrices, e.g. col / rowMedians(), col / rowRanks(),"| __truncated__
#>   .. .. .. ..$ License         : chr "Artistic-2.0"
#>   .. .. .. ..$ LazyLoad        : chr "TRUE"
#>   .. .. .. ..$ NeedsCompilation: chr "yes"
#>   .. .. .. ..$ ByteCompile     : chr "TRUE"
#>   .. .. .. ..$ URL             : chr "https://github.com/HenrikBengtsson/matrixStats"
#>   .. .. .. ..$ BugReports      : chr "https://github.com/HenrikBengtsson/matrixStats/issues"
#>   .. .. .. ..$ RoxygenNote     : chr "7.1.2"
#>   .. .. .. ..$ Packaged        : chr "2021-09-15 17:47:41 UTC; hb"
#>   .. .. .. ..$ Repository      : chr "CRAN"
#>   .. .. .. ..$ Date/Publication: chr "2021-09-17 11:10:02 UTC"
#>   .. .. .. ..$ Built           : chr "R 4.1.1; x86_64-pc-linux-gnu; 'Tue, 21 Sep 2021 10:27:37 +0200'; unix"
#>   .. .. .. ..- attr(*, "class")= chr "packageDescription"
#>   .. .. .. ..- attr(*, "file")= chr "/usr/lib/R/site-library/matrixStats/Meta/package.rds"
#>   .. ..$ matprod        : chr "default"
#>   .. ..$ BLAS           : chr "/usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3"
#>   .. ..$ LAPACK         : chr "/usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so"
#>   .. ..$ system.codepage: chr "NULL"
#>   .. ..$ codepage       : chr "NULL"
#>   .. ..- attr(*, "class")= chr "sessionInfo"
#>  $ allConfigurations:'data.frame':	148 obs. of  13 variables:
#>   ..$ .ID.       : int [1:148] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ algorithm  : chr [1:148] "as" "acs" "eas" "ras" ...
#>   ..$ localsearch: chr [1:148] "0" "3" "3" "1" ...
#>   ..$ alpha      : num [1:148] 1 1.13 3.72 4.84 4.21 ...
#>   ..$ beta       : num [1:148] 1 3.35 5.52 5.93 5.23 ...
#>   ..$ rho        : num [1:148] 0.95 0.647 0.6 0.761 0.761 ...
#>   ..$ ants       : num [1:148] 10 36 75 5 22 16 40 15 67 7 ...
#>   ..$ nnls       : num [1:148] NA 43 29 40 8 26 8 11 28 27 ...
#>   ..$ q0         : num [1:148] NA 0.905 NA NA NA ...
#>   ..$ dlb        : chr [1:148] NA "0" "1" "0" ...
#>   ..$ rasrank    : num [1:148] NA NA NA 61 49 NA NA NA NA NA ...
#>   ..$ elitistants: num [1:148] NA NA 91 NA NA NA NA 336 350 NA ...
#>   ..$ .PARENT.   : int [1:148] NA NA NA NA NA NA NA NA NA NA ...
#>  $ softRestart      : logi [1:7] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ iterationElites  : int [1:7] 2 50 50 111 123 119 123
#>  $ testing          :List of 2
#>   ..$ experiments: num [1:10, 1:16] 23609115 23327035 23111373 23128507 23340533 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "1t" "2t" "3t" "4t" ...
#>   .. .. ..$ : chr [1:16] "2" "29" "3" "20" ...
#>   ..$ seeds      : Named int [1:10] 1385446146 1396979195 448912041 732530909 947545849 195435663 798649446 1692971486 896414478 1371337352
#>   .. ..- attr(*, "names")= chr [1:10] "1t" "2t" "3t" "4t" ...