Read the log file produced by irace (irace.Rdata).

Read the log file produced by irace (irace.Rdata).

Usage

read_logfile(filename, name = "iraceResults")

Arguments

filename

Filename that contains the log file saved by irace. Example: irace.Rdata.

name

Optional argument that allows overriding the default name of the object in the file.

Value

(list())

Examples

irace_results <- read_logfile(system.file("exdata/irace-acotsp.Rdata", package="irace",
                                          mustWork=TRUE))
str(irace_results)
#> List of 11
#>  $ scenario         :List of 56
#>   ..$ scenarioFile             : chr "/home/manu/work/irace/git/devel-examples/vignette-example/scenario.txt"
#>   ..$ configurationsFile       : chr "/home/manu/work/irace/git/devel-examples/vignette-example/default.txt"
#>   ..$ logFile                  : chr "./irace-acotsp.Rdata"
#>   ..$ trainInstancesDir        : chr "./instances"
#>   ..$ trainInstancesFile       : chr "./instances.txt"
#>   ..$ testInstancesDir         : chr "./instances"
#>   ..$ testInstancesFile        : chr "./instances-test.txt"
#>   ..$ testNbElites             : int 5
#>   ..$ testIterationElites      : logi TRUE
#>   ..$ maxExperiments           : int 1000
#>   ..$ seed                     : int 687542627
#>   ..$ execDir                  : chr "./"
#>   ..$ parameterFile            : chr "./parameters.txt"
#>   ..$ recoveryFile             : chr ""
#>   ..$ instances                : chr [1:10] "./instances/1000-1.tsp" "./instances/1000-2.tsp" "./instances/1000-3.tsp" "./instances/1000-4.tsp" ...
#>   ..$ sampleInstances          : logi TRUE
#>   ..$ testInstances            : chr [1:10] "./instances/1000-10.tsp" "./instances/1000-11.tsp" "./instances/1000-12.tsp" "./instances/1000-13.tsp" ...
#>   ..$ testType                 : chr "friedman"
#>   ..$ firstTest                : int 5
#>   ..$ blockSize                : int 1
#>   ..$ eachTest                 : int 1
#>   ..$ targetRunner             : chr "./target-runner"
#>   ..$ targetRunnerLauncher     : chr ""
#>   ..$ targetCmdline            : chr "{configurationID} {instanceID} {seed} {instance} {bound} {targetRunnerArgs}"
#>   ..$ targetRunnerRetries      : int 0
#>   ..$ targetRunnerData         : chr ""
#>   ..$ deterministic            : logi FALSE
#>   ..$ maxTime                  : int 0
#>   ..$ budgetEstimation         : num 0.02
#>   ..$ minMeasurableTime        : num 0.01
#>   ..$ parallel                 : int 2
#>   ..$ loadBalancing            : logi TRUE
#>   ..$ mpi                      : logi FALSE
#>   ..$ batchmode                : chr "0"
#>   ..$ digits                   : int 4
#>   ..$ quiet                    : logi FALSE
#>   ..$ debugLevel               : int 0
#>   ..$ softRestart              : logi TRUE
#>   ..$ softRestartThreshold     : num 1e-04
#>   ..$ elitist                  : logi TRUE
#>   ..$ elitistNewInstances      : int 1
#>   ..$ elitistLimit             : int 2
#>   ..$ capping                  : logi FALSE
#>   ..$ cappingType              : chr "median"
#>   ..$ boundType                : chr "candidate"
#>   ..$ boundDigits              : int 0
#>   ..$ boundPar                 : int 1
#>   ..$ boundAsTimeout           : logi TRUE
#>   ..$ postselection            : num 0
#>   ..$ aclib                    : logi FALSE
#>   ..$ nbIterations             : int 0
#>   ..$ nbExperimentsPerIteration: int 0
#>   ..$ minNbSurvival            : int 0
#>   ..$ nbConfigurations         : int 0
#>   ..$ mu                       : int 5
#>   ..$ confidence               : num 0.95
#>  $ irace.version    : chr "3.5.1.22532d5"
#>  $ parameters       :List of 14
#>   ..$ names       : chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ types       : Named chr [1:11] "c" "c" "r" "r" ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ switches    : Named chr [1:11] "--" "--localsearch " "--alpha " "--beta " ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ domain      :List of 11
#>   .. ..$ algorithm  : chr [1:5] "as" "mmas" "eas" "ras" ...
#>   .. ..$ localsearch: chr [1:4] "0" "1" "2" "3"
#>   .. ..$ alpha      : num [1:2] 0 5
#>   .. ..$ beta       : num [1:2] 0 10
#>   .. ..$ rho        : num [1:2] 0.01 1
#>   .. ..$ ants       : num [1:2] 5 100
#>   .. ..$ nnls       : num [1:2] 5 50
#>   .. ..$ q0         : num [1:2] 0 1
#>   .. ..$ dlb        : chr [1:2] "0" "1"
#>   .. ..$ rasrank    : num [1:2] 1 100
#>   .. ..$ elitistants: num [1:2] 1 750
#>   ..$ conditions  :List of 11
#>   .. ..$ algorithm  : logi TRUE
#>   .. ..$ localsearch: logi TRUE
#>   .. ..$ alpha      : logi TRUE
#>   .. ..$ beta       : logi TRUE
#>   .. ..$ rho        : logi TRUE
#>   .. ..$ ants       : logi TRUE
#>   .. ..$ nnls       :  expression(localsearch %in% c(1, 2, 3))
#>   .. ..$ q0         :  expression(algorithm %in% c("acs"))
#>   .. ..$ dlb        :  expression(localsearch %in% c(1, 2, 3))
#>   .. ..$ rasrank    :  expression(algorithm %in% c("ras"))
#>   .. ..$ elitistants:  expression(algorithm %in% c("eas"))
#>   ..$ isFixed     : Named logi [1:11] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ transform   :List of 11
#>   .. ..$ algorithm  : chr ""
#>   .. ..$ localsearch: chr ""
#>   .. ..$ alpha      : chr ""
#>   .. ..$ beta       : chr ""
#>   .. ..$ rho        : chr ""
#>   .. ..$ ants       : chr "log"
#>   .. .. ..- attr(*, "lower")= num 1.61
#>   .. .. ..- attr(*, "upper")= num 4.62
#>   .. ..$ nnls       : chr ""
#>   .. ..$ q0         : chr ""
#>   .. ..$ dlb        : chr ""
#>   .. ..$ rasrank    : chr ""
#>   .. ..$ elitistants: chr ""
#>   ..$ isDependent : Named logi [1:11] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ digits      : Named int [1:4] 4 4 4 4
#>   .. ..- attr(*, "names")= chr [1:4] "alpha" "beta" "rho" "q0"
#>   ..$ depends     :List of 11
#>   .. ..$ algorithm  : chr(0) 
#>   .. ..$ localsearch: chr(0) 
#>   .. ..$ alpha      : chr(0) 
#>   .. ..$ beta       : chr(0) 
#>   .. ..$ rho        : chr(0) 
#>   .. ..$ ants       : chr(0) 
#>   .. ..$ nnls       : chr "localsearch"
#>   .. ..$ q0         : chr "algorithm"
#>   .. ..$ dlb        : chr "localsearch"
#>   .. ..$ rasrank    : chr "algorithm"
#>   .. ..$ elitistants: chr "algorithm"
#>   ..$ hierarchy   : Named num [1:11] 1 1 1 1 1 1 2 2 2 2 ...
#>   .. ..- attr(*, "names")= chr [1:11] "algorithm" "localsearch" "alpha" "beta" ...
#>   ..$ nbParameters: int 11
#>   ..$ nbFixed     : int 0
#>   ..$ nbVariable  : int 11
#>  $ allElites        :List of 7
#>   ..$ : num [1:4] 2 29 3 20
#>   ..$ : num [1:5] 50 2 38 29 34
#>   ..$ : num [1:5] 86 50 88 64 73
#>   ..$ : num 108
#>   ..$ : num [1:4] 108 127 136 145
#>   ..$ : num [1:5] 149 151 127 147 145
#>   ..$ : num [1:5] 149 151 147 127 145
#>  $ experiments      : num [1:15, 1:163] 28407696 28402347 29166659 28417102 28823388 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:15] "1" "2" "3" "4" ...
#>   .. ..$ : chr [1:163] "1" "2" "3" "4" ...
#>  $ experimentLog    : num [1:1000, 1:5] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "iteration" "instance" "configuration" "time" ...
#>  $ state            :List of 17
#>   ..$ .Random.seed        : int [1:626] 10403 176 74450096 -488718086 1971160917 1271940195 -1466109272 593734952 -700000720 -249115685 ...
#>   ..$ .irace              :<environment: 0x560f0faeb420> 
#>   ..$ currentBudget       : num 25
#>   ..$ debugLevel          : int 0
#>   ..$ eliteConfigurations :'data.frame':	5 obs. of  16 variables:
#>   .. ..$ .ID.       : num [1:5] 149 151 147 127 145
#>   .. ..$ algorithm  : chr [1:5] "acs" "acs" "acs" "acs" ...
#>   .. ..$ localsearch: chr [1:5] "3" "3" "3" "3" ...
#>   .. ..$ alpha      : num [1:5] 1.563 1.361 1.64 0.993 1.884
#>   .. ..$ beta       : num [1:5] 1.1449 5.3681 3.8003 4.4442 0.0659
#>   .. ..$ rho        : num [1:5] 0.269 0.345 0.441 0.459 0.239
#>   .. ..$ ants       : num [1:5] 8 5 10 7 7
#>   .. ..$ nnls       : num [1:5] 14 16 21 21 18
#>   .. ..$ q0         : num [1:5] 0.36 0.513 0.624 0.486 0.346
#>   .. ..$ dlb        : chr [1:5] "1" "1" "1" "1" ...
#>   .. ..$ rasrank    : num [1:5] NA NA NA NA NA
#>   .. ..$ elitistants: num [1:5] NA NA NA NA NA
#>   .. ..$ .PARENT.   : num [1:5] 145 127 127 108 108
#>   .. ..$ .ALIVE.    : logi [1:5] TRUE TRUE TRUE TRUE TRUE
#>   .. ..$ .RANK.     : num [1:5] 49 49 50 53 54
#>   .. ..$ .WEIGHT.   : num [1:5] 0.3333 0.2667 0.2 0.1333 0.0667
#>   ..$ experimentsUsedSoFar: int 1000
#>   ..$ indexIteration      : num 8
#>   ..$ minSurvival         : num 5
#>   ..$ model               :List of 11
#>   .. ..$ algorithm  :List of 5
#>   .. .. ..$ 149: num [1:5] 0.000235 0.000235 0.000235 0.005177 0.994119
#>   .. .. ..$ 151: num [1:5] 0.000235 0.000235 0.000235 0.005177 0.994119
#>   .. .. ..$ 127: num [1:5] 0.000235 0.000235 0.000235 0.005177 0.994119
#>   .. .. ..$ 147: num [1:5] 0.000235 0.000235 0.000235 0.005177 0.994119
#>   .. .. ..$ 145: num [1:5] 0.000235 0.000235 0.000235 0.005177 0.994119
#>   .. ..$ localsearch:List of 5
#>   .. .. ..$ 149: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 151: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 127: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 147: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. .. ..$ 145: num [1:4] 0.000341 0.000341 0.000341 0.998978
#>   .. ..$ alpha      :List of 5
#>   .. .. ..$ 149: num [1:2] 0.577 1.563
#>   .. .. ..$ 151: num [1:2] 0.577 1.361
#>   .. .. ..$ 127: num [1:2] 0.577 0.993
#>   .. .. ..$ 147: num [1:2] 0.577 1.64
#>   .. .. ..$ 145: num [1:2] 0.577 1.884
#>   .. ..$ beta       :List of 5
#>   .. .. ..$ 149: num [1:2] 1.15 1.14
#>   .. .. ..$ 151: num [1:2] 1.15 5.37
#>   .. .. ..$ 127: num [1:2] 1.15 4.44
#>   .. .. ..$ 147: num [1:2] 1.15 3.8
#>   .. .. ..$ 145: num [1:2] 1.1532 0.0659
#>   .. ..$ rho        :List of 5
#>   .. .. ..$ 149: num [1:2] 0.114 0.269
#>   .. .. ..$ 151: num [1:2] 0.114 0.345
#>   .. .. ..$ 127: num [1:2] 0.114 0.459
#>   .. .. ..$ 147: num [1:2] 0.114 0.441
#>   .. .. ..$ 145: num [1:2] 0.114 0.239
#>   .. ..$ ants       :List of 5
#>   .. .. ..$ 149: num [1:2] 0.115 8
#>   .. .. ..$ 151: num [1:2] 0.115 5
#>   .. .. ..$ 127: num [1:2] 0.115 7
#>   .. .. ..$ 147: num [1:2] 0.115 10
#>   .. .. ..$ 145: num [1:2] 0.115 7
#>   .. ..$ nnls       :List of 5
#>   .. .. ..$ 149: num [1:2] 5.19 14
#>   .. .. ..$ 151: num [1:2] 5.19 16
#>   .. .. ..$ 127: num [1:2] 5.19 21
#>   .. .. ..$ 147: num [1:2] 5.19 21
#>   .. .. ..$ 145: num [1:2] 5.19 18
#>   .. ..$ q0         :List of 5
#>   .. .. ..$ 149: num [1:2] 0.115 0.36
#>   .. .. ..$ 151: num [1:2] 0.115 0.513
#>   .. .. ..$ 127: num [1:2] 0.115 0.486
#>   .. .. ..$ 147: num [1:2] 0.115 0.624
#>   .. .. ..$ 145: num [1:2] 0.115 0.346
#>   .. ..$ dlb        :List of 5
#>   .. .. ..$ 149: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 151: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 127: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 147: num [1:2] 0.00102 0.99898
#>   .. .. ..$ 145: num [1:2] 0.00102 0.99898
#>   .. ..$ rasrank    :List of 5
#>   .. .. ..$ 149: num [1:2] 11.4 12
#>   .. .. ..$ 151: num [1:2] 11.4 12
#>   .. .. ..$ 127: num [1:2] 11.4 12
#>   .. .. ..$ 147: num [1:2] 11.4 12
#>   .. .. ..$ 145: num [1:2] 11.4 12
#>   .. ..$ elitistants:List of 5
#>   .. .. ..$ 149: num [1:2] 86.4 NA
#>   .. .. ..$ 151: num [1:2] 86.4 NA
#>   .. .. ..$ 127: num [1:2] 86.4 NA
#>   .. .. ..$ 147: num [1:2] 86.4 NA
#>   .. .. ..$ 145: num [1:2] 86.4 NA
#>   ..$ nbConfigurations    : num 6
#>   ..$ nbIterations        : num 7
#>   ..$ remainingBudget     : int 0
#>   ..$ timeUsed            : num 0
#>   ..$ boundEstimate       : logi NA
#>   ..$ rejectedIDs         : NULL
#>   ..$ completed           : chr "Budget exhausted"
#>   ..$ elapsed             : Named num [1:3] 10059 44.7 5274.2
#>   .. ..- attr(*, "names")= chr [1:3] "user" "system" "wallclock"
#>  $ allConfigurations:'data.frame':	163 obs. of  13 variables:
#>   ..$ .ID.       : num [1:163] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ algorithm  : chr [1:163] "as" "acs" "eas" "ras" ...
#>   ..$ localsearch: chr [1:163] "0" "3" "3" "1" ...
#>   ..$ alpha      : num [1:163] 1 1.13 3.72 4.84 4.21 ...
#>   ..$ beta       : num [1:163] 1 3.35 5.52 5.93 5.23 ...
#>   ..$ rho        : num [1:163] 0.95 0.647 0.6 0.761 0.761 ...
#>   ..$ ants       : num [1:163] 10 36 75 5 22 16 40 15 67 7 ...
#>   ..$ nnls       : num [1:163] NA 43 29 40 8 26 8 11 28 27 ...
#>   ..$ q0         : num [1:163] NA 0.905 NA NA NA ...
#>   ..$ dlb        : chr [1:163] NA "0" "1" "0" ...
#>   ..$ rasrank    : num [1:163] NA NA NA 61 49 NA NA NA NA NA ...
#>   ..$ elitistants: num [1:163] NA NA 91 NA NA NA NA 336 350 NA ...
#>   ..$ .PARENT.   : num [1:163] NA NA NA NA NA NA NA NA NA NA ...
#>  $ softRestart      : logi [1:7] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ iterationElites  : num [1:7] 2 50 86 108 108 149 149
#>  $ testing          :List of 2
#>   ..$ experiments: num [1:10, 1:18] 23454147 23272450 23038445 23113076 23249037 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:10] "1t" "2t" "3t" "4t" ...
#>   .. .. ..$ : chr [1:18] "2" "29" "3" "20" ...
#>   ..$ seeds      : Named int [1:10] 1475363490 1986478844 635764558 2018927617 1736421458 1828952358 1530093410 1928089170 680468367 471709667
#>   .. ..- attr(*, "names")= chr [1:10] "1t" "2t" "3t" "4t" ...